@@ -34,60 +34,108 @@
     "url": "https://github.com/tediousjs/tedious.git"
   },
   "engines": {
-    "node": ">= 4"
+    "node": ">= 6"
   },
   "publishConfig": {
     "tag": "next"
   },
   "dependencies": {
-    "babel-runtime": "^6.23.0",
-    "big-number": "0.3.1",
-    "bl": "^1.2.0",
-    "iconv-lite": "^0.4.11",
-    "readable-stream": "^2.2.6",
-    "sprintf": "0.1.5",
-    "sspi-client": "^0.1.0"
+    "@azure/ms-rest-nodeauth": "2.0.2",
+    "@js-joda/core": "^2.0.0",
+    "bl": "^3.0.0",
+    "depd": "^2.0.0",
+    "iconv-lite": "^0.5.0",
+    "jsbi": "^3.1.1",
+    "native-duplexpair": "^1.0.0",
+    "punycode": "^2.1.0",
+    "readable-stream": "^3.5.0",
+    "sprintf-js": "^1.1.2"
   },
   "devDependencies": {
-    "async": "^1.4.0",
-    "babel-cli": "^6.24.0",
-    "babel-eslint": "^7.2.1",
-    "babel-plugin-transform-runtime": "^6.23.0",
-    "babel-preset-env": "^1.2.2",
-    "babel-register": "^6.24.0",
-    "benchmark": "^2.1.0",
-    "eslint": "^3.19.0",
-    "mitm": "^1.3.2",
-    "nodeunit": "^0.11.0",
-    "rimraf": "^2.6.1",
-    "semantic-release": "^6.3.2",
-    "sinon": "^1.17.5",
-    "source-map-support": "^0.4.14"
+    "@babel/cli": "^7.8.3",
+    "@babel/core": "^7.8.3",
+    "@babel/node": "^7.8.3",
+    "@babel/plugin-proposal-class-properties": "^7.8.3",
+    "@babel/preset-env": "^7.8.3",
+    "@babel/preset-typescript": "^7.8.3",
+    "@babel/register": "^7.8.3",
+    "@commitlint/cli": "^8.3.5",
+    "@commitlint/config-conventional": "^8.3.4",
+    "@commitlint/travis-cli": "^8.3.5",
+    "@types/bl": "^2.1.0",
+    "@types/depd": "^1.1.32",
+    "@types/node": "^13.5.1",
+    "@types/readable-stream": "^2.3.5",
+    "@types/sprintf-js": "^1.1.2",
+    "@typescript-eslint/eslint-plugin": "^2.18.0",
+    "@typescript-eslint/parser": "^2.18.0",
+    "async": "^2.6.3",
+    "babel-plugin-istanbul": "^5.2.0",
+    "chai": "^4.1.2",
+    "codecov": "^3.6.1",
+    "eslint": "^5.16.0",
+    "mitm": "^1.7.0",
+    "mocha": "^6.2.2",
+    "nyc": "^14.1.1",
+    "rimraf": "^3.0.0",
+    "semantic-release": "^17.0.1",
+    "sinon": "^7.5.0",
+    "typescript": "^3.7.5"
   },
   "scripts": {
-    "pretest": "eslint src test",
-    "test": "nodeunit --reporter minimal test/setup.js test/unit/ test/unit/token/ test/unit/tracking-buffer",
-    "pretest-all": "eslint src test",
-    "test-all": "nodeunit --reporter minimal test/setup.js test/unit/ test/unit/token/ test/unit/tracking-buffer test/integration/",
-    "pretest-integration": "eslint src test",
-    "test-integration": "nodeunit --reporter minimal test/setup.js test/integration/",
-    "build": "rimraf lib && babel src --out-dir lib",
+    "lint": "eslint src test --ext .js,.ts && tsc",
+    "test": "mocha test/unit test/unit/token test/unit/tracking-buffer",
+    "test-integration": "mocha test/integration/",
+    "test-all": "mocha test/unit/ test/unit/token/ test/unit/tracking-buffer test/integration/",
+    "build": "rimraf lib && babel src --out-dir lib --extensions .js,.ts",
     "prepublish": "npm run build",
-    "semantic-release": "semantic-release pre && npm publish && semantic-release post"
+    "semantic-release": "semantic-release"
   },
   "babel": {
+    "ignore": [
+      "./src/**/*.d.ts"
+    ],
     "presets": [
       [
-        "env",
+        "@babel/preset-env",
         {
           "targets": {
-            "node": 4
+            "node": 6
           }
+        }
+      ],
+      [
+        "@babel/preset-typescript",
+        {
+          "allowDeclareFields": true
         }
       ]
     ],
     "plugins": [
-      "transform-runtime"
+      [
+        "@babel/transform-typescript",
+        {
+          "allowDeclareFields": true
+        }
+      ],
+      [
+        "@babel/plugin-proposal-class-properties",
+        {
+          "loose": true
+        }
+      ]
+    ]
+  },
+  "commitlint": {
+    "extends": [
+      "@commitlint/config-conventional"
+    ]
+  },
+  "nyc": {
+    "sourceMap": false,
+    "instrument": false,
+    "extension": [
+      ".ts"
     ]
   }
 }
