@@ -1,310 +1,356 @@
 const InstanceLookup = require('../../src/instance-lookup').InstanceLookup;
-const Sinon = require('sinon');
+const sinon = require('sinon');
+const dns = require('dns');
+const punycode = require('punycode');
+const assert = require('chai').assert;
 
-exports['instanceLookup invalid args'] = {
-  setUp: function(done) {
-    this.instanceLookup = new InstanceLookup().instanceLookup;
-    done();
-  },
+describe('instanceLookup invalid args', function() {
+  let instanceLookup;
 
-  'invalid server': function(test) {
-    const expectedErrorMessage = 'Invalid arguments: "server" must be a string';
-    try {
-      const notString = 4;
-      this.instanceLookup({ server: notString });
-    } catch (err) {
-      test.strictEqual(err.message, expectedErrorMessage);
-      test.done();
-    }
-  },
+  beforeEach(function() {
+    instanceLookup = new InstanceLookup();
+  });
 
-  'invalid instanceName': function(test) {
-    const expectedErrorMessage =
-      'Invalid arguments: "instanceName" must be a string';
-    try {
-      const notString = 4;
-      this.instanceLookup({ server: 'serverName', instanceName: notString });
-    } catch (err) {
-      test.strictEqual(err.message, expectedErrorMessage);
-      test.done();
-    }
-  },
+  it('invalid server', () => {
+    assert.throws(() => {
+      instanceLookup.instanceLookup({ server: 4 });
+    }, 'Invalid arguments: "server" must be a string');
+  });
 
-  'invalid timeout': function(test) {
-    const expectedErrorMessage =
-      'Invalid arguments: "timeout" must be a number';
-    try {
-      const notNumber = 'some string';
-      this.instanceLookup({
+  it('invalid instanceName', () => {
+    assert.throws(() => {
+      instanceLookup.instanceLookup({ server: 'serverName', instanceName: 4 });
+    }, 'Invalid arguments: "instanceName" must be a string');
+  });
+
+  it('invalid timeout', () => {
+    assert.throws(() => {
+      instanceLookup.instanceLookup({
         server: 'server',
         instanceName: 'instance',
-        timeout: notNumber
+        timeout: 'some string'
       });
-    } catch (err) {
-      test.strictEqual(err.message, expectedErrorMessage);
-      test.done();
-    }
-  },
+    }, 'Invalid arguments: "timeout" must be a number');
+  });
 
-  'invalid retries': function(test) {
-    const expectedErrorMessage =
-      'Invalid arguments: "retries" must be a number';
-    try {
-      const notNumber = 'some string';
-      this.instanceLookup({
+  it('invalid retries', () => {
+    assert.throws(() => {
+      instanceLookup.instanceLookup({
         server: 'server',
         instanceName: 'instance',
         timeout: 1000,
-        retries: notNumber
+        retries: 'some string'
       });
-    } catch (err) {
-      test.strictEqual(err.message, expectedErrorMessage);
-      test.done();
-    }
-  },
+    }, 'Invalid arguments: "retries" must be a number');
+  });
 
-  'invalid callback': function(test) {
-    const expectedErrorMessage =
-      'Invalid arguments: "callback" must be a function';
-    try {
-      const notFunction = 4;
-      this.instanceLookup(
-        {
-          server: 'server',
-          instanceName: 'instance',
-          timeout: 1000,
-          retries: 3
-        },
-        notFunction
-      );
-    } catch (err) {
-      test.strictEqual(err.message, expectedErrorMessage);
-      test.done();
-    }
-  }
-};
+  it('invalid callback', () => {
+    assert.throws(() => {
+      instanceLookup.instanceLookup({
+        server: 'server',
+        instanceName: 'instance',
+        timeout: 1000,
+        retries: 3
+      }, 4);
+    }, 'Invalid arguments: "callback" must be a function');
+  });
+});
 
-exports['instanceLookup functional unit tests'] = {
-  setUp: function(done) {
-    this.sinon = Sinon.sandbox.create();
+describe('instanceLookup functional unit tests', function() {
 
-    this.options = {
+  let options;
+  let anyPort;
+  let anyRequest;
+  let anyMessage;
+  let anyError;
+  let anySqlPort;
+  let instanceLookup;
+  let testSender;
+  let createSenderStub;
+  let senderExecuteStub;
+  let parseStub;
+
+
+  beforeEach(function() {
+    options = {
       server: 'server',
       instanceName: 'instance',
       timeout: 1000,
       retries: 3
     };
 
-    this.anyPort = 1234;
-    this.anyRequest = new Buffer(0x02);
-    this.anyMessage = 'any message';
-    this.anyError = new Error('any error');
-    this.anySqlPort = 2345;
+    anyPort = 1234;
+    anyRequest = Buffer.alloc(0x02);
+    anyMessage = 'any message';
+    anyError = new Error('any error');
+    anySqlPort = 2345;
 
-    this.instanceLookup = new InstanceLookup();
+    instanceLookup = new InstanceLookup();
 
     // Stub out createSender method to return the Sender we create. This allows us
     // to override the execute method on Sender so we can test instance lookup code
     // without triggering network activity.
-    this.testSender = this.instanceLookup.createSender(
-      this.options.server,
-      this.anyPort,
-      this.anyRequest
+    testSender = instanceLookup.createSender(
+      options.server,
+      anyPort,
+      anyRequest
     );
-    this.createSenderStub = this.sinon.stub(
-      this.instanceLookup,
+    createSenderStub = sinon.stub(
+      instanceLookup,
       'createSender'
     );
-    this.createSenderStub.returns(this.testSender);
-    this.senderExecuteStub = this.sinon.stub(this.testSender, 'execute');
+    createSenderStub.returns(testSender);
+    senderExecuteStub = sinon.stub(testSender, 'execute');
 
     // Stub parseBrowserResponse so we can mimic success and failure without creating
     // elaborate responses. parseBrowserResponse itself has unit tests to ensure that
     // it functions correctly.
-    this.parseStub = this.sinon.stub(
-      this.instanceLookup,
+    parseStub = sinon.stub(
+      instanceLookup,
       'parseBrowserResponse'
     );
+  });
 
-    done();
-  },
+  afterEach(function() {
+    sinon.restore();
+  }),
 
-  tearDown: function(done) {
-    this.sinon.restore();
-    done();
-  },
+  it('success', (done) => {
+    senderExecuteStub.callsArgWithAsync(0, null, anyMessage);
+    parseStub
+      .withArgs(anyMessage, options.instanceName)
+      .returns(anySqlPort);
 
-  success: function(test) {
-    this.senderExecuteStub.callsArgWithAsync(0, null, this.anyMessage);
-    this.parseStub
-      .withArgs(this.anyMessage, this.options.instanceName)
-      .returns(this.anySqlPort);
+    instanceLookup.instanceLookup(options, (error, port) => {
+      assert.strictEqual(error, undefined);
+      assert.strictEqual(port, anySqlPort);
 
-    this.instanceLookup.instanceLookup(this.options, (error, port) => {
-      test.strictEqual(error, undefined);
-      test.strictEqual(port, this.anySqlPort);
+      assert.ok(createSenderStub.calledOnce);
+      assert.strictEqual(createSenderStub.args[0][0], options.server);
+      assert.strictEqual(createSenderStub.args[0][3]);
 
-      test.ok(this.createSenderStub.calledOnce);
-      test.strictEqual(this.createSenderStub.args[0][0], this.options.server);
-      test.strictEqual(this.createSenderStub.args[0][3]);
+      assert.ok(senderExecuteStub.calledOnce);
+      assert.ok(parseStub.calledOnce);
 
-      test.ok(this.senderExecuteStub.calledOnce);
-      test.ok(this.parseStub.calledOnce);
-      test.done();
+      done();
     });
-  },
+  }),
 
-  'sender fail': function(test) {
-    this.senderExecuteStub.callsArgWithAsync(0, this.anyError, undefined);
+  it('sender fail', (done) => {
+    senderExecuteStub.callsArgWithAsync(0, anyError, undefined);
 
-    this.instanceLookup.instanceLookup(this.options, (error, port) => {
-      test.ok(error.indexOf(this.anyError.message) !== -1);
-      test.strictEqual(port, undefined);
+    instanceLookup.instanceLookup(options, (error, port) => {
+      assert.ok(error.indexOf(anyError.message) !== -1);
+      assert.strictEqual(port, undefined);
 
-      test.ok(this.createSenderStub.calledOnce);
-      test.strictEqual(this.createSenderStub.args[0][0], this.options.server);
-      test.strictEqual(this.createSenderStub.args[0][3]);
+      assert.ok(createSenderStub.calledOnce);
+      assert.strictEqual(createSenderStub.args[0][0], options.server);
+      assert.strictEqual(createSenderStub.args[0][3]);
 
-      test.ok(this.senderExecuteStub.calledOnce);
-      test.strictEqual(this.parseStub.callCount, 0);
-      test.done();
+      assert.ok(senderExecuteStub.calledOnce);
+      assert.strictEqual(parseStub.callCount, 0);
+
+      done();
     });
-  },
+  }),
 
-  'parse fail': function(test) {
-    this.senderExecuteStub.callsArgWithAsync(0, null, this.anyMessage);
-    this.parseStub
-      .withArgs(this.anyMessage, this.options.instanceName)
+  it('parse fail', (done) => {
+    senderExecuteStub.callsArgWithAsync(0, null, anyMessage);
+    parseStub
+      .withArgs(anyMessage, options.instanceName)
       .returns(null);
 
-    this.instanceLookup.instanceLookup(this.options, (error, port) => {
-      test.ok(error.indexOf('not found') !== -1);
-      test.strictEqual(port, undefined);
+    instanceLookup.instanceLookup(options, (error, port) => {
+      assert.ok(error.indexOf('not found') !== -1);
+      assert.strictEqual(port, undefined);
 
-      test.ok(this.createSenderStub.calledOnce);
-      test.strictEqual(this.createSenderStub.args[0][0], this.options.server);
-      test.strictEqual(this.createSenderStub.args[0][3]);
+      assert.ok(createSenderStub.calledOnce);
+      assert.strictEqual(createSenderStub.args[0][0], options.server);
+      assert.strictEqual(createSenderStub.args[0][3]);
 
-      test.ok(this.senderExecuteStub.calledOnce);
-      test.ok(this.parseStub.calledOnce);
-      test.done();
+      assert.ok(senderExecuteStub.calledOnce);
+      assert.ok(parseStub.calledOnce);
+
+      done();
     });
-  },
+  }),
 
-  'retry success': function(test) {
+  it('retry success', (done) => {
     // First invocation of execute will not invoke callback. This will cause a timeout
     // and trigger a retry. Setup to invoke callback on second invocation.
-    this.senderExecuteStub
+    senderExecuteStub
       .onCall(1)
-      .callsArgWithAsync(0, null, this.anyMessage);
-    this.parseStub
-      .withArgs(this.anyMessage, this.options.instanceName)
-      .returns(this.anySqlPort);
+      .callsArgWithAsync(0, null, anyMessage);
+    parseStub
+      .withArgs(anyMessage, options.instanceName)
+      .returns(anySqlPort);
 
-    const clock = this.sinon.useFakeTimers();
+    const clock = sinon.useFakeTimers();
 
-    this.instanceLookup.instanceLookup(this.options, (error, port) => {
-      test.strictEqual(error, undefined);
-      test.strictEqual(port, this.anySqlPort);
+    instanceLookup.instanceLookup(options, (error, port) => {
+      assert.strictEqual(error, undefined);
+      assert.strictEqual(port, anySqlPort);
 
-      test.ok(this.createSenderStub.callCount, 2);
-      for (let j = 0; j < this.createSenderStub.callCount; j++) {
-        test.strictEqual(this.createSenderStub.args[j][0], this.options.server);
-        test.strictEqual(this.createSenderStub.args[j][3]);
+      assert.ok(createSenderStub.callCount, 2);
+      for (let j = 0; j < createSenderStub.callCount; j++) {
+        assert.strictEqual(createSenderStub.args[j][0], options.server);
+        assert.strictEqual(createSenderStub.args[j][3]);
       }
 
       // Execute called twice but parse only called once as the first call to execute times out.
-      test.strictEqual(this.senderExecuteStub.callCount, 2);
-      test.ok(this.parseStub.calledOnce);
+      assert.strictEqual(senderExecuteStub.callCount, 2);
+      assert.ok(parseStub.calledOnce);
 
       clock.restore();
-      test.done();
+
+      done();
     });
 
     // Forward clock to trigger timeout.
-    clock.tick(this.options.timeout * 1.1);
-  },
+    clock.tick(options.timeout * 1.1);
+  }),
 
-  'retry fail': function(test) {
-    const clock = this.sinon.useFakeTimers();
+  it('retry fail', (done) => {
+    const clock = sinon.useFakeTimers();
 
     const forwardClock = () => {
-      clock.tick(this.options.timeout * 1.1);
+      clock.tick(options.timeout * 1.1);
     };
 
-    const scheduleForwardClock = function() {
+    function scheduleForwardClock() {
       // This function is called in place of sender.execute(). We don't want to
       // rely on when the timeout is set in relation to execute() in the calling
       // context. So we setup to forward clock on next tick.
       process.nextTick(forwardClock);
-    };
+    }
 
-    this.senderExecuteStub.restore();
-    this.senderExecuteStub = this.sinon.stub(
-      this.testSender,
+    senderExecuteStub.restore();
+    senderExecuteStub = sinon.stub(
+      testSender,
       'execute',
-      scheduleForwardClock
-    );
+    ).callsFake(scheduleForwardClock);
 
-    this.instanceLookup.instanceLookup(this.options, (error, port) => {
-      test.ok(error.indexOf('Failed to get response') != -1);
-      test.strictEqual(port, undefined);
+    instanceLookup.instanceLookup(options, (error, port) => {
+      assert.ok(error.indexOf('Failed to get response') !== -1);
+      assert.strictEqual(port, undefined);
 
-      test.strictEqual(this.createSenderStub.callCount, this.options.retries);
-      for (let j = 0; j < this.createSenderStub.callCount; j++) {
-        test.strictEqual(this.createSenderStub.args[j][0], this.options.server);
-        test.strictEqual(this.createSenderStub.args[j][3]);
+      assert.strictEqual(createSenderStub.callCount, options.retries);
+      for (let j = 0; j < createSenderStub.callCount; j++) {
+        assert.strictEqual(createSenderStub.args[j][0], options.server);
+        assert.strictEqual(createSenderStub.args[j][3]);
       }
 
       // Execute called 'retries' number of times but parse is never called because
       // all the execute calls timeout.
-      test.strictEqual(this.senderExecuteStub.callCount, this.options.retries);
-      test.strictEqual(this.parseStub.callCount, 0);
+      assert.strictEqual(senderExecuteStub.callCount, options.retries);
+      assert.strictEqual(parseStub.callCount, 0);
 
       clock.restore();
-      test.done();
+
+      done();
     });
-  }
-};
+  }),
 
-exports['parseBrowserResponse'] = {
-  setUp: function(done) {
-    this.parse = new InstanceLookup().parseBrowserResponse;
-    done();
-  },
+  it('incorrect instanceName', (done) => {
+    const message = 'ServerName;WINDOWS2;InstanceName;XXXXXXXXXX;IsClustered;No;Version;10.50.2500.0;tcp;0;;' +
+      'ServerName;WINDOWS2;InstanceName;YYYYYYYYYY;IsClustered;No;Version;10.50.2500.0;tcp;0;;';
+    senderExecuteStub.callsArgWithAsync(0, null, message);
+    parseStub
+      .withArgs(message, options.instanceName);
 
-  oneInstanceFound: function(test) {
+    instanceLookup.instanceLookup(options, (error, port) => {
+      assert.ok(error.indexOf('XXXXXXXXXX') === -1);
+      assert.ok(error.indexOf('YYYYYYYYYY') === -1);
+      assert.strictEqual(port, undefined);
+
+      assert.ok(createSenderStub.calledOnce);
+      assert.ok(senderExecuteStub.calledOnce);
+      assert.ok(parseStub.calledOnce);
+
+      done();
+    });
+  });
+});
+
+describe('parseBrowserResponse', function() {
+  let instanceLookup;
+
+  beforeEach(function() {
+    instanceLookup = new InstanceLookup();
+  });
+
+  it('oneInstanceFound', () => {
     const response =
       'ServerName;WINDOWS2;InstanceName;SQLEXPRESS;IsClustered;No;Version;10.50.2500.0;tcp;1433;;';
 
-    test.strictEqual(this.parse(response, 'sqlexpress'), 1433);
-    test.done();
-  },
+    assert.strictEqual(instanceLookup.parseBrowserResponse(response, 'sqlexpress'), 1433);
+  });
 
-  twoInstancesFoundInFirst: function(test) {
+  it('twoInstancesFoundInFirst', () => {
     const response =
       'ServerName;WINDOWS2;InstanceName;SQLEXPRESS;IsClustered;No;Version;10.50.2500.0;tcp;1433;;' +
       'ServerName;WINDOWS2;InstanceName;XXXXXXXXXX;IsClustered;No;Version;10.50.2500.0;tcp;0;;';
 
-    test.strictEqual(this.parse(response, 'sqlexpress'), 1433);
-    test.done();
-  },
+    assert.strictEqual(instanceLookup.parseBrowserResponse(response, 'sqlexpress'), 1433);
+  });
 
-  twoInstancesFoundInSecond: function(test) {
+  it('twoInstancesFoundInSecond', () => {
     const response =
       'ServerName;WINDOWS2;InstanceName;XXXXXXXXXX;IsClustered;No;Version;10.50.2500.0;tcp;0;;' +
       'ServerName;WINDOWS2;InstanceName;SQLEXPRESS;IsClustered;No;Version;10.50.2500.0;tcp;1433;;';
 
-    test.strictEqual(this.parse(response, 'sqlexpress'), 1433);
-    test.done();
-  },
+    assert.strictEqual(instanceLookup.parseBrowserResponse(response, 'sqlexpress'), 1433);
+  });
 
-  twoInstancesNotFound: function(test) {
+  it('twoInstancesNotFound', () => {
     const response =
       'ServerName;WINDOWS2;InstanceName;XXXXXXXXXX;IsClustered;No;Version;10.50.2500.0;tcp;0;;' +
       'ServerName;WINDOWS2;InstanceName;YYYYYYYYYY;IsClustered;No;Version;10.50.2500.0;tcp;0;;';
 
-    test.strictEqual(this.parse(response, 'sqlexpress'), undefined);
-    test.done();
-  }
-};
+    assert.strictEqual(instanceLookup.parseBrowserResponse(response, 'sqlexpress'), undefined);
+  });
+});
+
+describe('parseBrowserResponse', function() {
+  let spy;
+
+  beforeEach(function() {
+    spy = sinon.spy(dns, 'lookup');
+  });
+
+  afterEach(function() {
+    sinon.restore();
+  });
+
+  it('test IDN Server name', (done) => {
+    const options = {
+      server: '本地主机.ad',
+      instanceName: 'instance',
+      timeout: 500,
+      retries: 1
+    };
+
+    new InstanceLookup().instanceLookup(options, () => {
+      assert.ok(spy.called, 'Failed to call dns.lookup on hostname');
+      assert.ok(spy.calledWithMatch(punycode.toASCII(options.server)), 'Unexpected hostname passed to dns.lookup');
+
+      done();
+    });
+  });
+
+  it('test ASCII Server name', (done) => {
+    const options = {
+      server: 'localhost',
+      instanceName: 'instance',
+      timeout: 500,
+      retries: 1
+    };
+
+    new InstanceLookup().instanceLookup(options, () => {
+      assert.ok(spy.called, 'Failed to call dns.lookup on hostname');
+      assert.ok(spy.calledWithMatch(options.server), 'Unexpected hostname passed to dns.lookup');
+
+      done();
+    });
+  });
+});
