@@ -1,91 +1,106 @@
 const Connection = require('../../src/tedious').Connection;
 const fs = require('fs');
-const Sinon = require('sinon');
+const sinon = require('sinon');
 const TransientErrorLookup = require('../../src/transient-error-lookup').TransientErrorLookup;
+const assert = require('chai').assert;
 
-const getConfig = function() {
-  const config = JSON.parse(fs.readFileSync(process.env.HOME + '/.tedious/test-connection.json', 'utf8')).config;
-  config.password = 'InvalidPassword';
+function getConfig() {
+  const config = JSON.parse(fs.readFileSync(require('os').homedir() + '/.tedious/test-connection.json', 'utf8')).config;
+  if (config.authentication) {
+    config.authentication.options.password = 'InvalidPassword';
+  } else {
+    config.password = 'InvalidPassword';
+  }
   config.options.maxRetriesOnTransientErrors = 5;
   config.options.connectionRetryInterval = 25;
 
   return config;
-};
+}
 
-exports['connection retry tests'] = {
-  setUp: function(done) {
-    this.invalidLoginError = 18456;
-    this.sinon = Sinon.sandbox.create();
+describe('Connection Retry Test', function() {
+  let invalidLoginError;
+
+  beforeEach(function(done) {
+    invalidLoginError = 18456;
     done();
-  },
+  });
 
-  tearDown: function(done) {
-    this.sinon.restore();
+  afterEach(function(done) {
+    sinon.restore();
     done();
-  },
+  });
 
-  'retry specified number of times on transient errors': function(test) {
+  it('should retry specified number of times on transient errors', function(done) {
     const config = getConfig();
-    test.expect(config.options.maxRetriesOnTransientErrors + 1);
 
-    this.sinon.stub(TransientErrorLookup.prototype, 'isTransientError', (error) => {
-      return error === this.invalidLoginError;
+    if (config.authentication && config.authentication.type === 'azure-active-directory-password') {
+      return done();
+    }
+
+    sinon.stub(TransientErrorLookup.prototype, 'isTransientError').callsFake((error) => {
+      return error === invalidLoginError;
     });
 
     const connection = new Connection(config);
 
     connection.on('retry', () => {
-      test.ok(true);
+      assert.ok(true);
     });
 
     connection.on('connect', (err) => {
-      test.ok(err);
+      assert.ok(err);
     });
 
     connection.on('end', (info) => {
-      test.done();
+      done();
     });
-  },
+  });
 
-  'no retries on non-transient errors': function(test) {
+  it('should no retries on non-transient errors', function(done) {
     const config = getConfig();
-    test.expect(1);
 
-    this.sinon.stub(TransientErrorLookup.prototype, 'isTransientError', (error) => {
-      return error !== this.invalidLoginError;
+    if (config.authentication && config.authentication.type === 'azure-active-directory-password') {
+      return done();
+    }
+
+    sinon.stub(TransientErrorLookup.prototype, 'isTransientError').callsFake((error) => {
+      return error !== invalidLoginError;
     });
 
     const connection = new Connection(config);
 
     connection.on('retry', () => {
-      test.ok(false);
+      assert.ok(false);
     });
 
     connection.on('connect', (err) => {
-      test.ok(err);
+      assert.ok(err);
     });
 
     connection.on('end', (info) => {
-      test.done();
+      done();
     });
-  },
+  });
 
-  'no retries if connection timeout fires': function(test) {
+  it('should no retries if connection timeout fires', function(done) {
     const config = getConfig();
+
+    if (config.authentication && config.authentication.type === 'azure-active-directory-password') {
+      return done();
+    }
+
     config.options.connectTimeout = config.options.connectionRetryInterval / 2;
 
-    const clock = this.sinon.useFakeTimers();
+    const clock = sinon.useFakeTimers({ toFake: ['setTimeout'] });
 
-    test.expect(1);
-
-    this.sinon.stub(TransientErrorLookup.prototype, 'isTransientError', (error) => {
-      return error === this.invalidLoginError;
+    sinon.stub(TransientErrorLookup.prototype, 'isTransientError').callsFake((error) => {
+      return error === invalidLoginError;
     });
 
     const connection = new Connection(config);
 
     connection.on('retry', () => {
-      test.ok(false);
+      assert.ok(false);
     });
 
     connection.on('errorMessage', () => {
@@ -94,12 +109,12 @@ exports['connection retry tests'] = {
     });
 
     connection.on('connect', (err) => {
-      test.ok(err);
+      assert.ok(err);
     });
 
     connection.on('end', (info) => {
       clock.restore();
-      test.done();
+      done();
     });
-  },
-};
+  });
+});
