@@ -1,11 +1,11 @@
-var Connection = require('../../src/connection');
-var Request = require('../../src/request');
-var fs = require('fs');
+const Connection = require('../../src/connection');
+const Request = require('../../src/request');
+const fs = require('fs');
+const assert = require('chai').assert;
+const debug = false;
 
-var debug = false;
-
-var config = JSON.parse(
-  fs.readFileSync(process.env.HOME + '/.tedious/test-connection.json', 'utf8')
+const config = JSON.parse(
+  fs.readFileSync(require('os').homedir() + '/.tedious/test-connection.json', 'utf8')
 ).config;
 config.options.textsize = 8 * 1024;
 
@@ -23,117 +23,24 @@ if (debug) {
 
 config.options.tdsVersion = process.env.TEDIOUS_TDS_VERSION;
 
-exports.uniqueConstraint = function(test) {
-  var sql = `\
-create table #testUnique (id int unique);
-insert #testUnique values (1), (2), (3);
-insert #testUnique values (2);
-drop table #testUnique;\
-`;
+function execSql(done, sql, requestCallback) {
+  const connection = new Connection(config);
 
-  test.expect(3);
-  execSql(test, sql, function(err) {
-    test.ok(err instanceof Error);
-    test.strictEqual(err.number, 2627);
-  });
-};
-
-exports.nullable = function(test) {
-  var sql = `\
-create table #testNullable (id int not null);
-insert #testNullable values (null);
-drop table #testNullable;\
-`;
-
-  test.expect(3);
-  execSql(test, sql, function(err) {
-    test.ok(err instanceof Error);
-    test.strictEqual(err.number, 515);
-  });
-};
-
-exports.cannotDropProcedure = function(test) {
-  var sql = '\
-drop procedure #nonexistentProcedure;\
-';
-
-  test.expect(3);
-  execSql(test, sql, function(err) {
-    test.ok(err instanceof Error);
-    test.strictEqual(err.number, 3701);
-  });
-};
-
-// Create a temporary stored procedure to test that err.procName,
-// err.lineNumber, err.class, and err.state are correct.
-//
-// We can't really test serverName reliably, other than that it exists.
-exports.extendedErrorInfo = function(test) {
-  var connection = new Connection(config);
-
-  test.expect(9);
-
-  var execProc = new Request('#testExtendedErrorInfo', function(err) {
-    test.ok(err instanceof Error);
-
-    test.strictEqual(err.number, 50000);
-    test.strictEqual(err.state, 42, 'err.state wrong');
-    test.strictEqual(err.class, 14, 'err.class wrong');
-
-    test.ok(err.serverName != null, 'err.serverName not set');
-
-    // The procedure name will actually be padded to 128 chars with underscores and
-    // some random hexadecimal digits.
-    test.ok(
-      (err.procName != null
-        ? err.procName.indexOf('#testExtendedErrorInfo')
-        : undefined) === 0,
-      `err.procName should begin with #testExtendedErrorInfo, was actually ${err.procName}`
-    );
-    test.strictEqual(err.lineNumber, 1, 'err.lineNumber should be 1');
-
-    connection.close();
-  });
-
-  var createProc = new Request(
-    "create procedure #testExtendedErrorInfo as raiserror('test error message', 14, 42)",
-    function(err) {
-      test.ifError(err);
-      connection.callProcedure(execProc);
-    }
-  );
-
-  connection.on('connect', function(err) {
-    test.ifError(err);
-    connection.execSqlBatch(createProc);
-  });
-
-  connection.on('end', function(info) {
-    test.done();
-  });
-
-  if (debug) {
-    connection.on('debug', function(message) {
-      console.log(message);
-    });
-  }
-};
-
-var execSql = function(test, sql, requestCallback) {
-  var connection = new Connection(config);
-
-  var request = new Request(sql, function() {
+  const request = new Request(sql, function() {
     requestCallback.apply(this, arguments);
     connection.close();
   });
 
   connection.on('connect', function(err) {
-    test.ifError(err);
+    if (err) {
+      return done(err);
+    }
+
     connection.execSqlBatch(request);
   });
 
   connection.on('end', function(info) {
-    test.done();
+    done();
   });
 
   if (debug) {
@@ -141,4 +48,128 @@ var execSql = function(test, sql, requestCallback) {
       console.log(message);
     });
   }
-};
+}
+
+describe('Errors Test', function() {
+  it('should test unique constraints', function(done) {
+    const sql = `\
+  create table #testUnique (id int unique);
+  insert #testUnique values (1), (2), (3);
+  insert #testUnique values (2);
+  drop table #testUnique;\
+  `;
+
+    execSql(done, sql, function(err) {
+      assert.ok(err instanceof Error);
+      assert.strictEqual(err.number, 2627);
+    });
+  });
+
+  it('should test nullabe', function(done) {
+    const sql = `\
+  create table #testNullable (id int not null);
+  insert #testNullable values (null);
+  drop table #testNullable;\
+  `;
+
+    execSql(done, sql, function(err) {
+      assert.ok(err instanceof Error);
+      assert.strictEqual(err.number, 515);
+    });
+  });
+
+  it('should test', function(done) {
+    const sql = '\
+  drop procedure #nonexistentProcedure;\
+  ';
+
+    execSql(done, sql, function(err) {
+      assert.ok(err instanceof Error);
+      assert.strictEqual(err.number, 3701);
+    });
+  });
+
+
+  // Create a temporary stored procedure to test that err.procName,
+  // err.lineNumber, err.class, and err.state are correct.
+  //
+  // We can't really test serverName reliably, other than that it exists.
+  it('should test extended error info', function(done) {
+    const connection = new Connection(config);
+
+    const execProc = new Request('#testExtendedErrorInfo', function(err) {
+      assert.ok(err instanceof Error);
+
+      assert.strictEqual(err.number, 50000);
+      assert.strictEqual(err.state, 42, 'err.state wrong');
+      assert.strictEqual(err.class, 14, 'err.class wrong');
+
+      assert.ok(err.serverName != null, 'err.serverName not set');
+
+      // The procedure name will actually be padded to 128 chars with underscores and
+      // some random hexadecimal digits.
+      assert.ok(
+        (err.procName != null ?
+          err.procName.indexOf('#testExtendedErrorInfo') :
+          undefined) === 0,
+        `err.procName should begin with #testExtendedErrorInfo, was actually ${err.procName}`
+      );
+      assert.strictEqual(err.lineNumber, 1, 'err.lineNumber should be 1');
+
+      connection.close();
+    });
+
+    const createProc = new Request(
+      "create procedure #testExtendedErrorInfo as raiserror('test error message', 14, 42)",
+      function(err) {
+        if (err) {
+          return done(err);
+        }
+
+        connection.callProcedure(execProc);
+      }
+    );
+
+    connection.on('connect', function(err) {
+      if (err) {
+        return done(err);
+      }
+
+      connection.execSqlBatch(createProc);
+    });
+
+    connection.on('end', function(info) {
+      done();
+    });
+
+    if (debug) {
+      connection.on('debug', function(message) {
+        console.log(message);
+      });
+    }
+  });
+
+  it('should support cancelling after starting query execution', function(done) {
+    const connection = new Connection(config);
+
+    const request = new Request("select 42, 'hello world'", function(err, rowCount) {
+      if (err) {
+        assert.equal(err.message, 'Canceled.');
+      }
+      connection.close();
+    });
+
+    connection.on('connect', function(err) {
+      if (err) {
+        return done(err);
+      }
+
+      connection.execSql(request);
+      connection.cancel();
+    });
+
+    connection.on('end', function(info) {
+      done();
+    });
+  });
+});
