@@ -1,14 +1,14 @@
-var Connection = require('../../src/connection');
-var Request = require('../../src/request');
-var Transaction = require('../../src/transaction');
+const Connection = require('../../src/connection');
+const Request = require('../../src/request');
+const Transaction = require('../../src/transaction');
+const fs = require('fs');
+const async = require('async');
+const assert = require('chai').assert;
 
-var fs = require('fs');
-var async = require('async');
+const debug = false;
 
-var debug = false;
-
-var config = JSON.parse(
-  fs.readFileSync(process.env.HOME + '/.tedious/test-connection.json', 'utf8')
+const config = JSON.parse(
+  fs.readFileSync(require('os').homedir() + '/.tedious/test-connection.json', 'utf8')
 ).config;
 
 if (debug) {
@@ -26,7 +26,7 @@ if (debug) {
 config.options.tdsVersion = process.env.TEDIOUS_TDS_VERSION;
 
 class Tester {
-  constructor(test) {
+  constructor(done) {
     this.createTable = this.createTable.bind(this);
     this.createProc = this.createProc.bind(this);
     this.execProc = this.execProc.bind(this);
@@ -41,11 +41,11 @@ class Tester {
     this.rollbackTransaction = this.rollbackTransaction.bind(this);
     this.close = this.close.bind(this);
     this.run = this.run.bind(this);
-    this.test = test;
+    this.done = done;
     this.connection = new Connection(config);
 
     this.connection.on('end', (info) => {
-      this.test.done();
+      this.done();
     });
 
     this.connection.on('errorMessage', (error) => {
@@ -60,10 +60,10 @@ class Tester {
   }
 
   createTable(callback) {
-    var request = new Request(
+    const request = new Request(
       'create table #temp (id int)',
       function(err) {
-        this.test.ifError(err);
+        assert.ifError(err);
         callback(err);
       }.bind(this)
     );
@@ -72,7 +72,7 @@ class Tester {
   }
 
   createProc(callback) {
-    var request = new Request(
+    const request = new Request(
       `\
 CREATE PROCEDURE #proc
 AS
@@ -83,7 +83,7 @@ AS
   commit transaction
 GO`,
       function(err) {
-        this.test.ifError(err);
+        assert.ifError(err);
         callback(err);
       }.bind(this)
     );
@@ -92,10 +92,10 @@ GO`,
   }
 
   execProc(callback) {
-    var request = new Request(
+    const request = new Request(
       'exec #proc',
       function(err) {
-        this.test.ifError(err);
+        assert.ifError(err);
         callback(err);
       }.bind(this)
     );
@@ -104,10 +104,10 @@ GO`,
   }
 
   insert(callback) {
-    var request = new Request(
+    const request = new Request(
       'insert into #temp (id) values(1)',
       function(err) {
-        this.test.ifError(err);
+        assert.ifError(err);
         callback(err);
       }.bind(this)
     );
@@ -116,17 +116,17 @@ GO`,
   }
 
   select(callback, expectedRows) {
-    var request = new Request(
+    const request = new Request(
       'select id from #temp',
       function(err, rowCount) {
-        this.test.ifError(err);
-        this.test.strictEqual(rowCount, expectedRows);
+        assert.ifError(err);
+        assert.strictEqual(rowCount, expectedRows);
         callback(err);
       }.bind(this)
     );
 
     request.on('row', (columns) => {
-      this.test.strictEqual(columns[0].value, 1);
+      assert.strictEqual(columns[0].value, 1);
     });
 
     this.connection.execSqlBatch(request);
@@ -142,8 +142,8 @@ GO`,
 
   beginTransaction(callback, transactionName) {
     this.connection.beginTransaction((err, transactionDescriptor) => {
-      this.test.ifError(err);
-      this.test.ok(
+      assert.ifError(err);
+      assert.ok(
         config.options.tdsVersion < '7_2' ? true : transactionDescriptor
       );
 
@@ -161,7 +161,7 @@ GO`,
 
   commitTransaction(callback) {
     this.connection.commitTransaction((err) => {
-      this.test.ifError(err);
+      assert.ifError(err);
 
       callback(err);
     });
@@ -169,7 +169,7 @@ GO`,
 
   rollbackTransaction(callback) {
     this.connection.rollbackTransaction((err) => {
-      this.test.ifError(err);
+      assert.ifError(err);
 
       callback(err);
     });
@@ -186,493 +186,469 @@ GO`,
   }
 }
 
-exports.transactionCommit = function(test) {
-  test.expect(8);
+describe('Transactions Test', function() {
+  it('should test transaction commit', function(done) {
+    const tester = new Tester(done);
+    tester.run([
+      tester.createTable,
+      tester.beginTransaction1,
+      tester.insert,
+      tester.commitTransaction,
+      tester.selectExpectOneRow,
+      tester.close
+    ]);
+  });
 
-  var tester = new Tester(test);
-  tester.run([
-    tester.createTable,
-    tester.beginTransaction1,
-    tester.insert,
-    tester.commitTransaction,
-    tester.selectExpectOneRow,
-    tester.close
-  ]);
-};
+  it('should test transaction rollback', function(done) {
+    const tester = new Tester(done);
+    tester.run([
+      tester.createTable,
+      tester.beginTransaction1,
+      tester.insert,
+      tester.rollbackTransaction,
+      tester.selectExpectZeroRows,
+      tester.close
+    ]);
+  });
 
-exports.transactionRollback = function(test) {
-  test.expect(7);
+  it('should test nested transaction commit', function(done) {
+    const tester = new Tester(done);
+    tester.run([
+      tester.createTable,
+      tester.beginTransaction1,
+      tester.beginTransaction2,
+      tester.insert,
+      tester.commitTransaction,
+      tester.commitTransaction,
+      tester.selectExpectOneRow,
+      tester.close
+    ]);
+  });
 
-  var tester = new Tester(test);
-  tester.run([
-    tester.createTable,
-    tester.beginTransaction1,
-    tester.insert,
-    tester.rollbackTransaction,
-    tester.selectExpectZeroRows,
-    tester.close
-  ]);
-};
+  it('should test tested transaction rollback outer', function(done) {
+    const tester = new Tester(done);
+    tester.run([
+      tester.createTable,
+      tester.beginTransaction1,
+      tester.beginTransaction2,
+      tester.insert,
+      tester.commitTransaction,
+      tester.rollbackTransaction,
+      tester.selectExpectZeroRows,
+      tester.close
+    ]);
+  });
 
-exports.nestedTransactionCommit = function(test) {
-  test.expect(11);
+  it('should test nested transaction in proc commit', function(done) {
+    const tester = new Tester(done);
+    tester.run([
+      tester.createTable,
+      tester.createProc,
+      tester.beginTransaction1,
+      tester.execProc,
+      tester.commitTransaction,
+      tester.selectExpectOneRow,
+      tester.close
+    ]);
+  });
 
-  var tester = new Tester(test);
-  tester.run([
-    tester.createTable,
-    tester.beginTransaction1,
-    tester.beginTransaction2,
-    tester.insert,
-    tester.commitTransaction,
-    tester.commitTransaction,
-    tester.selectExpectOneRow,
-    tester.close
-  ]);
-};
+  it('should test nested transaction in proc rollback outer', function(done) {
+    const tester = new Tester(done);
+    tester.run([
+      tester.createTable,
+      tester.createProc,
+      tester.beginTransaction1,
+      tester.execProc,
+      tester.rollbackTransaction,
+      tester.selectExpectZeroRows,
+      tester.close
+    ]);
+  });
 
-exports.nestedTransactionRollbackOuter = function(test) {
-  test.expect(10);
+  it('should test first rollback transaction event with xact abort', function(done) {
+    // From 2.2.7.8, ENVCHANGE_TOKEN type Begin Transaction (8) is only supported
+    // in TDS version 7.2 and above. 'rollbackTransaction' event fires in response
+    // to that token type and hence won't be firing for lower versions.
+    if (config.options.tdsVersion < '7_2') {
+      // test.expect(4);
+    } else {
+      // test.expect(5);
+    }
 
-  var tester = new Tester(test);
-  tester.run([
-    tester.createTable,
-    tester.beginTransaction1,
-    tester.beginTransaction2,
-    tester.insert,
-    tester.commitTransaction,
-    tester.rollbackTransaction,
-    tester.selectExpectZeroRows,
-    tester.close
-  ]);
-};
+    const connection = new Connection(config);
+    connection.on('end', (info) => done());
+    //  connection.on('errorMessage', (error) => console.log("#{error.number} : #{error.message}"))
+    //  connection.on('debug', (message) => console.log(message))
 
-exports.nestedTransactionInProcCommit = function(test) {
-  test.expect(9);
+    connection.on('connect', function(err) {
+      let req = new Request('create table #temp (value varchar(50))', function(
+        err
+      ) {
+        assert.ifError(err);
 
-  var tester = new Tester(test);
-  tester.run([
-    tester.createTable,
-    tester.createProc,
-    tester.beginTransaction1,
-    tester.execProc,
-    tester.commitTransaction,
-    tester.selectExpectOneRow,
-    tester.close
-  ]);
-};
+        req = new Request('SET XACT_ABORT ON', function(err) {
+          assert.ifError(err);
 
-exports.nestedTransactionInProcRollbackOuter = function(test) {
-  test.expect(8);
+          connection.beginTransaction(function(err) {
+            assert.ifError(err);
 
-  var tester = new Tester(test);
-  tester.run([
-    tester.createTable,
-    tester.createProc,
-    tester.beginTransaction1,
-    tester.execProc,
-    tester.rollbackTransaction,
-    tester.selectExpectZeroRows,
-    tester.close
-  ]);
-};
+            connection.on('rollbackTransaction', function() {
+              // Ensure rollbackTransaction event is fired
+              assert.ok(true);
+            });
 
-exports.firesRollbackTransactionEventWithXactAbort = function(test) {
-  // From 2.2.7.8, ENVCHANGE_TOKEN type Begin Transaction (8) is only supported
-  // in TDS version 7.2 and above. 'rollbackTransaction' event fires in response
-  // to that token type and hence won't be firing for lower versions.
-  if (config.options.tdsVersion < '7_2') {
-    test.expect(4);
-  } else {
-    test.expect(5);
-  }
+            req = new Request(
+              "insert into #temp values ('asdfasdfasdfasdfasdfasdfasdfasdfasdfasdfasdfasdfasd')",
+              function(err) {
+                assert.strictEqual(
+                  err.message,
+                  'String or binary data would be truncated.'
+                );
 
-  var connection = new Connection(config);
-  connection.on('end', (info) => test.done());
-  //  connection.on('errorMessage', (error) => console.log("#{error.number} : #{error.message}"))
-  //  connection.on('debug', (message) => console.log(message))
-
-  connection.on('connect', function(err) {
-    var req = new Request('create table #temp (value varchar(50))', function(
-      err
-    ) {
-      test.ifError(err);
-
-      req = new Request('SET XACT_ABORT ON', function(err) {
-        test.ifError(err);
-
-        connection.beginTransaction(function(err) {
-          test.ifError(err);
-
-          connection.on('rollbackTransaction', function() {
-            // Ensure rollbackTransaction event is fired
-            test.ok(true);
+                connection.close();
+              }
+            );
+            connection.execSqlBatch(req);
           });
-
-          req = new Request(
-            "insert into #temp values ('asdfasdfasdfasdfasdfasdfasdfasdfasdfasdfasdfasdfasd')",
-            function(err) {
-              test.strictEqual(
-                err.message,
-                'String or binary data would be truncated.'
-              );
-
-              connection.close();
-            }
-          );
-          connection.execSqlBatch(req);
         });
+        connection.execSqlBatch(req);
       });
       connection.execSqlBatch(req);
     });
-    connection.execSqlBatch(req);
   });
-};
 
-exports.transactionHelper = function(test) {
-  test.expect(3);
+  it('should test transaction helper', function(done) {
+    const connection = new Connection(config);
+    connection.on('end', (info) => done());
+    //  connection.on('errorMessage', (error) => console.log("#{error.number} : #{error.message}"))
+    //  connection.on('debug', (message) => console.log(message) if (debug))
 
-  var connection = new Connection(config);
-  connection.on('end', (info) => test.done());
-  //  connection.on('errorMessage', (error) => console.log("#{error.number} : #{error.message}"))
-  //  connection.on('debug', (message) => console.log(message) if (debug))
+    connection.on('connect', function(err) {
+      connection.transaction(function(err, outerDone) {
+        assert.ifError(err);
 
-  connection.on('connect', function(err) {
-    connection.transaction(function(err, outerDone) {
-      test.ifError(err);
+        connection.transaction(function(err, innerDone) {
+          assert.ifError(err);
 
-      connection.transaction(function(err, innerDone) {
-        test.ifError(err);
-
-        innerDone(null, outerDone, function(err) {
-          test.ifError(err);
-          connection.close();
+          innerDone(null, outerDone, function(err) {
+            assert.ifError(err);
+            connection.close();
+          });
         });
       });
     });
   });
-};
 
-exports.transactionHelperSelectiveRollback = function(test) {
-  test.expect(9);
+  it('should test transaction helper selective rollback', function(done) {
+    const connection = new Connection(config);
+    connection.on('end', (info) => done());
+    //  connection.on('errorMessage', (error) => console.log("#{error.number} : #{error.message}"))
+    //  connection.on('debug', (message) => console.log(message) if (debug))
 
-  var connection = new Connection(config);
-  connection.on('end', (info) => test.done());
-  //  connection.on('errorMessage', (error) => console.log("#{error.number} : #{error.message}"))
-  //  connection.on('debug', (message) => console.log(message) if (debug))
+    connection.on('connect', function(err) {
+      let request = new Request('create table #temp (id int)', function(err) {
+        assert.ifError(err);
 
-  connection.on('connect', function(err) {
-    var request = new Request('create table #temp (id int)', function(err) {
-      test.ifError(err);
+        connection.transaction(function(err, outerDone) {
+          assert.ifError(err);
 
-      connection.transaction(function(err, outerDone) {
-        test.ifError(err);
+          request = new Request('insert into #temp (id) VALUES (1)', function(
+            err
+          ) {
+            assert.ifError(err);
 
-        request = new Request('insert into #temp (id) VALUES (1)', function(
-          err
-        ) {
-          test.ifError(err);
+            connection.transaction(function(err, innerDone) {
+              assert.ifError(err);
 
-          connection.transaction(function(err, innerDone) {
-            test.ifError(err);
+              request = new Request('insert into #temp (id) VALUES (2)', function(
+                err
+              ) {
+                assert.ifError(err);
 
-            request = new Request('insert into #temp (id) VALUES (2)', function(
-              err
-            ) {
-              test.ifError(err);
+                const expectedError = new Error('Something failed');
+                innerDone(expectedError, function(err) {
+                  assert.strictEqual(err, expectedError);
 
-              var expectedError = new Error('Something failed');
-              innerDone(expectedError, function(err) {
-                test.strictEqual(err, expectedError);
+                  // Do not pass the error to the outer transaction continuation
+                  outerDone(null, function(err) {
+                    assert.ifError(err);
 
-                // Do not pass the error to the outer transaction continuation
-                outerDone(null, function(err) {
-                  test.ifError(err);
+                    request = new Request('select * from #temp', function(err) {
+                      assert.ifError(err);
+                      connection.close();
+                    });
+
+                    request.on('row', function(row) {
+                      assert.strictEqual(row[0].value, 1);
+                    });
+
+                    connection.execSql(request);
+                  });
+                });
+              });
+
+              connection.execSql(request);
+            });
+          });
+          connection.execSql(request);
+        });
+      });
+      connection.execSqlBatch(request);
+    });
+  });
+
+  it('should test transaction helper full rollback', function(done) {
+    const connection = new Connection(config);
+    connection.on('end', (info) => done());
+    //  connection.on('errorMessage', (error) => console.log("#{error.number} : #{error.message}"))
+    //  connection.on('debug', (message) => console.log(message) if (debug))
+
+    connection.on('connect', function(err) {
+      let request = new Request('create table #temp (id int)', function(err) {
+        assert.ifError(err);
+
+        connection.transaction(function(err, outerDone) {
+          assert.ifError(err);
+
+          request = new Request('insert into #temp (id) VALUES (1)', function(
+            err
+          ) {
+            assert.ifError(err);
+
+            connection.transaction(function(err, innerDone) {
+              assert.ifError(err);
+
+              request = new Request('insert into #temp (id) VALUES (2)', function(
+                err
+              ) {
+                assert.ifError(err);
+
+                const expectedError = new Error('Something failed');
+                innerDone(expectedError, outerDone, function(err) {
+                  assert.strictEqual(err, expectedError);
 
                   request = new Request('select * from #temp', function(err) {
-                    test.ifError(err);
+                    assert.ifError(err);
                     connection.close();
                   });
 
                   request.on('row', function(row) {
-                    test.strictEqual(row[0].value, 1);
+                    throw new Error('Did not expect any rows');
                   });
 
                   connection.execSql(request);
                 });
               });
-            });
 
-            connection.execSql(request);
-          });
-        });
-        connection.execSql(request);
-      });
-    });
-    connection.execSqlBatch(request);
-  });
-};
-
-exports.transactionHelperFullRollback = function(test) {
-  test.expect(7);
-
-  var connection = new Connection(config);
-  connection.on('end', (info) => test.done());
-  //  connection.on('errorMessage', (error) => console.log("#{error.number} : #{error.message}"))
-  //  connection.on('debug', (message) => console.log(message) if (debug))
-
-  connection.on('connect', function(err) {
-    var request = new Request('create table #temp (id int)', function(err) {
-      test.ifError(err);
-
-      connection.transaction(function(err, outerDone) {
-        test.ifError(err);
-
-        request = new Request('insert into #temp (id) VALUES (1)', function(
-          err
-        ) {
-          test.ifError(err);
-
-          connection.transaction(function(err, innerDone) {
-            test.ifError(err);
-
-            request = new Request('insert into #temp (id) VALUES (2)', function(
-              err
-            ) {
-              test.ifError(err);
-
-              var expectedError = new Error('Something failed');
-              innerDone(expectedError, outerDone, function(err) {
-                test.strictEqual(err, expectedError);
-
-                request = new Request('select * from #temp', function(err) {
-                  test.ifError(err);
-                  connection.close();
-                });
-
-                request.on('row', function(row) {
-                  throw new Error('Did not expect any rows');
-                });
-
-                connection.execSql(request);
-              });
-            });
-
-            connection.execSql(request);
-          });
-        });
-        connection.execSql(request);
-      });
-    });
-    connection.execSqlBatch(request);
-  });
-};
-
-exports.transactionHelperBatchAbortingError = function(test) {
-  test.expect(4);
-
-  var connection = new Connection(config);
-  connection.on('end', (info) => test.done());
-  //  connection.on('errorMessage', (error) => console.log("#{error.number} : #{error.message}"))
-  //  connection.on('debug', (message) => console.log(message) if (debug))
-
-  connection.on('connect', function(err) {
-    connection.transaction(function(err, outerDone) {
-      test.ifError(err);
-
-      connection.transaction(function(err, innerDone) {
-        test.ifError(err);
-
-        var request = new Request('create table #temp (id int)', function(err) {
-          test.ifError(err);
-
-          request = new Request('create table #temp (id int)', function(err) {
-            innerDone(err, outerDone, function(err) {
-              test.equal(
-                err.message,
-                "There is already an object named '#temp' in the database."
-              );
-
-              connection.close();
+              connection.execSql(request);
             });
           });
-
-          connection.execSqlBatch(request);
+          connection.execSql(request);
         });
-        connection.execSqlBatch(request);
       });
-    });
-  });
-};
-
-exports.transactionHelperSocketError = function(test) {
-  test.expect(3);
-
-  var connection = new Connection(config);
-  connection.on('end', function(info) {
-    test.done();
-  });
-  connection.on('error', function(err) {
-    test.ok(~err.message.indexOf('socket error'));
-  });
-  //  connection.on('errorMessage', (error) => console.log("#{error.number} : #{error.message}"))
-  //  connection.on('debug', (message) => console.log(message) if (debug))
-
-  connection.on('connect', function(err) {
-    connection.transaction(function(err, outerDone) {
-      test.ifError(err);
-
-      connection.transaction(function(err, innerDone) {
-        test.ifError(err);
-
-        var request = new Request('WAITFOR 00:00:30', function(err) {
-          innerDone(err, outerDone, function(err) {
-            test.ok(~err.message.indexOf('socket error'));
-          });
-        });
-
-        connection.execSql(request);
-        connection.socket.emit('error', new Error('socket error'));
-      });
-    });
-  });
-};
-
-exports.transactionHelperIsolationLevel = function(test) {
-  test.expect(8);
-
-  var connection = new Connection(config);
-  connection.on('end', (info) => test.done());
-  //  connection.on('errorMessage', (error) => console.log("#{error.number} : #{error.message}"))
-  //  connection.on('debug', (message) => console.log(message) if (debug))
-
-  connection.on('connect', function(err) {
-    connection.transaction(function(err, outerDone) {
-      test.ifError(err);
-
-      var request = new Request(
-        'SELECT [transaction_isolation_level] FROM [sys].[dm_exec_sessions] WHERE [session_id] = @@SPID',
-        function(err) {
-          test.ifError(err);
-
-          connection.transaction(function(err, innerDone) {
-            test.ifError(err);
-
-            request = new Request(
-              'SELECT [transaction_isolation_level] FROM [sys].[dm_exec_sessions] WHERE [session_id] = @@SPID',
-              function(err) {
-                test.ifError(err);
-
-                innerDone(null, outerDone, function(err) {
-                  request = new Request(
-                    'SELECT [transaction_isolation_level] FROM [sys].[dm_exec_sessions] WHERE [session_id] = @@SPID',
-                    function(err) {
-                      test.ifError(err);
-
-                      connection.close();
-                    }
-                  );
-
-                  request.on('row', function(row) {
-                    test.equal(
-                      row[0].value,
-                      Transaction.ISOLATION_LEVEL.SERIALIZABLE
-                    );
-                  });
-
-                  connection.execSqlBatch(request);
-                });
-              }
-            );
-
-            request.on('row', function(row) {
-              test.equal(
-                row[0].value,
-                Transaction.ISOLATION_LEVEL.SERIALIZABLE
-              );
-            });
-
-            connection.execSqlBatch(request);
-          }, Transaction.ISOLATION_LEVEL.SERIALIZABLE);
-        }
-      );
-
-      request.on('row', function(row) {
-        test.equal(
-          row[0].value,
-          Transaction.ISOLATION_LEVEL.REPEATABLE_READ
-        );
-      });
-
       connection.execSqlBatch(request);
-    }, Transaction.ISOLATION_LEVEL.REPEATABLE_READ);
-  });
-};
-
-exports.transactionHelperResetOpenTransactionCount = function(test) {
-  test.expect(3);
-
-  var connection = new Connection(config);
-  connection.on('end', (info) => test.done());
-  //  connection.on('errorMessage', (error) => console.log("#{error.number} : #{error.message}"))
-  //  connection.on('debug', (message) => console.log(message) if (debug))
-
-  connection.on('connect', function(err) {
-    connection.transaction(function(err) {
-      test.ifError(err);
-
-      connection.reset(function(err) {
-        test.ifError(err);
-
-        test.strictEqual(connection.inTransaction, false);
-        connection.close();
-      });
     });
   });
-};
 
-exports.transactionHelperMixedWithLowLevelTransactionMethods = function(test) {
-  test.expect(11);
+  it('should test transaction helper batch aborting error', function(done) {
+    const connection = new Connection(config);
+    connection.on('end', (info) => done());
+    //  connection.on('errorMessage', (error) => console.log("#{error.number} : #{error.message}"))
+    //  connection.on('debug', (message) => console.log(message) if (debug))
 
-  var connection = new Connection(config);
-  connection.on('end', (info) => test.done());
-  //  connection.on('errorMessage', (error) => console.log("#{error.number} : #{error.message}"))
-  //  connection.on('debug', (message) => console.log(message) if (debug))
+    connection.on('connect', function(err) {
+      connection.transaction(function(err, outerDone) {
+        assert.ifError(err);
 
-  connection.on('connect', function(err) {
-    connection.beginTransaction(function(err) {
-      test.ifError(err);
+        connection.transaction(function(err, innerDone) {
+          assert.ifError(err);
 
-      test.strictEqual(connection.inTransaction, true);
+          let request = new Request('create table #temp (id int)', function(err) {
+            assert.ifError(err);
 
-      connection.transaction(function(err, txDone) {
-        test.ifError(err);
-
-        test.strictEqual(connection.inTransaction, true);
-
-        connection.beginTransaction(function(err) {
-          test.ifError(err);
-
-          test.strictEqual(connection.inTransaction, true);
-
-          connection.commitTransaction(function(err) {
-            test.ifError(err);
-
-            test.strictEqual(connection.inTransaction, true);
-
-            txDone(null, function(err) {
-              test.strictEqual(connection.inTransaction, true);
-
-              connection.commitTransaction(function(err) {
-                test.ifError(err);
-
-                test.strictEqual(connection.inTransaction, false);
+            request = new Request('create table #temp (id int)', function(err) {
+              innerDone(err, outerDone, function(err) {
+                assert.equal(
+                  err.message,
+                  "There is already an object named '#temp' in the database."
+                );
 
                 connection.close();
               });
             });
+
+            connection.execSqlBatch(request);
+          });
+          connection.execSqlBatch(request);
+        });
+      });
+    });
+  });
+
+  it('should test transaction helper socket error', function(done) {
+    const connection = new Connection(config);
+    connection.on('end', function(info) {
+      done();
+    });
+    connection.on('error', function(err) {
+      assert.ok(~err.message.indexOf('socket error'));
+    });
+    //  connection.on('errorMessage', (error) => console.log("#{error.number} : #{error.message}"))
+    //  connection.on('debug', (message) => console.log(message) if (debug))
+
+    connection.on('connect', function(err) {
+      connection.transaction(function(err, outerDone) {
+        assert.ifError(err);
+
+        connection.transaction(function(err, innerDone) {
+          assert.ifError(err);
+
+          const request = new Request('WAITFOR 00:00:30', function(err) {
+            assert.ok(~err.message.indexOf('socket error'));
+
+            innerDone(err, outerDone, function(err) {
+              assert.ok(~err.message.indexOf('socket error'));
+            });
+          });
+
+          connection.execSql(request);
+          connection.socket.emit('error', new Error('socket error'));
+        });
+      });
+    });
+  });
+
+  it('should test transaction helper isolation level', function(done) {
+    const connection = new Connection(config);
+    connection.on('end', (info) => done());
+    //  connection.on('errorMessage', (error) => console.log("#{error.number} : #{error.message}"))
+    //  connection.on('debug', (message) => console.log(message) if (debug))
+
+    connection.on('connect', function(err) {
+      connection.transaction(function(err, outerDone) {
+        assert.ifError(err);
+
+        let request = new Request(
+          'SELECT [transaction_isolation_level] FROM [sys].[dm_exec_sessions] WHERE [session_id] = @@SPID',
+          function(err) {
+            assert.ifError(err);
+
+            connection.transaction(function(err, innerDone) {
+              assert.ifError(err);
+
+              request = new Request(
+                'SELECT [transaction_isolation_level] FROM [sys].[dm_exec_sessions] WHERE [session_id] = @@SPID',
+                function(err) {
+                  assert.ifError(err);
+
+                  innerDone(null, outerDone, function(err) {
+                    request = new Request(
+                      'SELECT [transaction_isolation_level] FROM [sys].[dm_exec_sessions] WHERE [session_id] = @@SPID',
+                      function(err) {
+                        assert.ifError(err);
+
+                        connection.close();
+                      }
+                    );
+
+                    request.on('row', function(row) {
+                      assert.equal(
+                        row[0].value,
+                        Transaction.ISOLATION_LEVEL.SERIALIZABLE
+                      );
+                    });
+
+                    connection.execSqlBatch(request);
+                  });
+                }
+              );
+
+              request.on('row', function(row) {
+                assert.equal(
+                  row[0].value,
+                  Transaction.ISOLATION_LEVEL.SERIALIZABLE
+                );
+              });
+
+              connection.execSqlBatch(request);
+            }, Transaction.ISOLATION_LEVEL.SERIALIZABLE);
+          }
+        );
+
+        request.on('row', function(row) {
+          assert.equal(
+            row[0].value,
+            Transaction.ISOLATION_LEVEL.REPEATABLE_READ
+          );
+        });
+
+        connection.execSqlBatch(request);
+      }, Transaction.ISOLATION_LEVEL.REPEATABLE_READ);
+    });
+  });
+
+  it('should test transaction helper reset open transaction count', function(done) {
+    const connection = new Connection(config);
+    connection.on('end', (info) => done());
+    //  connection.on('errorMessage', (error) => console.log("#{error.number} : #{error.message}"))
+    //  connection.on('debug', (message) => console.log(message) if (debug))
+
+    connection.on('connect', function(err) {
+      connection.transaction(function(err) {
+        assert.ifError(err);
+
+        connection.reset(function(err) {
+          assert.ifError(err);
+
+          assert.strictEqual(connection.inTransaction, false);
+          connection.close();
+        });
+      });
+    });
+  });
+
+  it('should test transaction helper mixed with low level transaction methods', function(done) {
+    const connection = new Connection(config);
+    connection.on('end', (info) => done());
+    //  connection.on('errorMessage', (error) => console.log("#{error.number} : #{error.message}"))
+    //  connection.on('debug', (message) => console.log(message) if (debug))
+
+    connection.on('connect', function(err) {
+      connection.beginTransaction(function(err) {
+        assert.ifError(err);
+
+        assert.strictEqual(connection.inTransaction, true);
+
+        connection.transaction(function(err, txDone) {
+          assert.ifError(err);
+
+          assert.strictEqual(connection.inTransaction, true);
+
+          connection.beginTransaction(function(err) {
+            assert.ifError(err);
+
+            assert.strictEqual(connection.inTransaction, true);
+
+            connection.commitTransaction(function(err) {
+              assert.ifError(err);
+
+              assert.strictEqual(connection.inTransaction, true);
+
+              txDone(null, function(err) {
+                assert.strictEqual(connection.inTransaction, true);
+
+                connection.commitTransaction(function(err) {
+                  assert.ifError(err);
+
+                  assert.strictEqual(connection.inTransaction, false);
+
+                  connection.close();
+                });
+              });
+            });
           });
         });
       });
     });
   });
-};
+});
