@@ -1,21 +1,18 @@
-var Connection = require('../../src/connection');
-var Request = require('../../src/request');
-var TYPES = require('../../src/tedious').TYPES;
-var fs = require('fs');
+const Connection = require('../../src/connection');
+const Request = require('../../src/request');
+const TYPES = require('../../src/tedious').TYPES;
+const fs = require('fs');
+const async = require('async');
 
-var config = JSON.parse(
-  fs.readFileSync(process.env.HOME + '/.tedious/test-connection.json', 'utf8')
-).config;
-config.options.tdsVersion = process.env.TEDIOUS_TDS_VERSION;
+const { assert } = require('chai');
 
-var TEST_SETUP_1 =
-  'BEGIN TRY DROP PROCEDURE __tediousTvpTest DROP TYPE TediousTestType END TRY BEGIN CATCH END CATCH';
-var TEST_SETUP_2 =
-  'CREATE TYPE TediousTestType AS TABLE (a bit, b tinyint, c smallint, d int, e bigint, f real, g float, h varchar (100), i nvarchar (100), j datetime);';
-var TEST_SETUP_3 =
-  'CREATE PROCEDURE __tediousTvpTest (@tvp TediousTestType readonly) AS BEGIN select * from @tvp END';
+function getConfig() {
+  var config = JSON.parse(
+    fs.readFileSync(require('os').homedir() + '/.tedious/test-connection.json', 'utf8')
+  ).config;
 
-var getConfig = function() {
+  config.options.tdsVersion = process.env.TEDIOUS_TDS_VERSION;
+
   config.options.debug = {
     packet: true,
     data: true,
@@ -25,136 +22,145 @@ var getConfig = function() {
   };
 
   return config;
-};
+}
 
-exports.callProcedureWithTVP = function(test) {
-  config = getConfig();
-  if (config.options.tdsVersion < '7_3_A') {
-    console.log(`TVP is not supported on TDS ${config.options.tdsVersion}.`);
-    test.done();
-    return;
-  }
+describe('calling a procedure that takes and returns a TVP', function() {
+  this.timeout(5000);
 
-  test.expect(13);
+  let connection;
 
-  var request = new Request(TEST_SETUP_1, function(err, rowCount) {
-    connection.execSqlBatch(request2);
+  beforeEach(function(done) {
+    const config = getConfig();
+
+    if (config.options.tdsVersion < '7_3_A') {
+      this.skip();
+    }
+
+    connection = new Connection(config);
+
+    connection.on(
+      'infoMessage',
+      function(info) {}
+      // console.log("#{info.number} : #{info.message}")
+    );
+
+    connection.on(
+      'debug',
+      function(text) {}
+      // console.log(text)
+    );
+
+    connection.on('connect', done);
   });
 
-  var request2 = new Request(TEST_SETUP_2, function(err, rowCount) {
-    connection.execSqlBatch(request3);
-  });
-
-  var request3 = new Request(TEST_SETUP_3, function(err, rowCount) {
-    connection.callProcedure(request4);
-  });
-
-  var request4 = new Request('__tediousTvpTest', function(err, rowCount) {
+  afterEach(function() {
     connection.close();
   });
 
-  request4.on('doneInProc', function(rowCount, more) {
-    test.ok(more);
-    test.strictEqual(rowCount, 1);
-  });
-
-  request4.on('row', function(columns) {
-    test.strictEqual(columns[0].value, false);
-    test.strictEqual(columns[1].value, 1);
-    test.strictEqual(columns[2].value, 2);
-    test.strictEqual(columns[3].value, 3);
-    test.strictEqual(columns[4].value, '4');
-    test.strictEqual(columns[5].value, 5.5);
-    test.strictEqual(columns[6].value, 6.6);
-    test.strictEqual(columns[7].value, 'asdf');
-    test.strictEqual(columns[8].value, 'asdf');
-    test.strictEqual(+columns[9].value, +new Date(Date.UTC(2014, 0, 1)));
-  });
-
-  var table = {
-    columns: [
-      {
-        name: 'a',
-        type: TYPES.Bit
+  it('returns the same data', function(done) {
+    async.series([
+      (next) => {
+        const sql = 'USE tempdb; BEGIN TRY DROP TYPE TediousTestType END TRY BEGIN CATCH END CATCH';
+        connection.execSqlBatch(new Request(sql, next));
       },
-      {
-        name: 'b',
-        type: TYPES.TinyInt
+      (next) => {
+        const sql = 'USE tempdb; CREATE TYPE TediousTestType AS TABLE (a bit, b tinyint, c smallint, d int, e bigint, f real, g float, h varchar (100), i nvarchar (100), j datetime)';
+        connection.execSqlBatch(new Request(sql, next));
       },
-      {
-        name: 'c',
-        type: TYPES.SmallInt
-      },
-      {
-        name: 'd',
-        type: TYPES.Int
-      },
-      {
-        name: 'e',
-        type: TYPES.BigInt
-      },
-      {
-        name: 'f',
-        type: TYPES.Real
-      },
-      {
-        name: 'g',
-        type: TYPES.Float
-      },
-      {
-        name: 'h',
-        type: TYPES.VarChar,
-        length: 100
-      },
-      {
-        name: 'i',
-        type: TYPES.NVarChar,
-        length: 100
-      },
-      {
-        name: 'j',
-        type: TYPES.DateTime,
-        length: 100
+      (next) => {
+        const sql = 'CREATE PROCEDURE #__tediousTvpTest @tvp TediousTestType readonly AS BEGIN select * from @tvp END';
+        connection.execSqlBatch(new Request(sql, next));
       }
-    ],
-    rows: [
-      [
-        false,
-        1,
-        2,
-        3,
-        4,
-        5.5,
-        6.6,
-        'asdf',
-        'asdf',
-        new Date(Date.UTC(2014, 0, 1))
-      ]
-    ]
-  };
+    ], (err) => {
+      if (err) {
+        console.log(err.message);
+        return done();
+      }
 
-  request4.addParameter('tvp', TYPES.TVP, table, {});
+      var request4 = new Request('#__tediousTvpTest', done);
 
-  var connection = new Connection(config);
+      request4.on('doneInProc', function(rowCount, more) {
+        assert.strictEqual(rowCount, 1);
+      });
 
-  connection.on('connect', function(err) {
-    test.ifError(err);
-    connection.execSqlBatch(request);
+      request4.on('row', function(columns) {
+        assert.strictEqual(columns[0].value, false);
+        assert.strictEqual(columns[1].value, 1);
+        assert.strictEqual(columns[2].value, 2);
+        assert.strictEqual(columns[3].value, 3);
+        assert.strictEqual(columns[4].value, '4');
+        assert.strictEqual(columns[5].value, 5.5);
+        assert.strictEqual(columns[6].value, 6.6);
+        assert.strictEqual(columns[7].value, 'asdf');
+        assert.strictEqual(columns[8].value, 'asdf');
+        assert.strictEqual(+columns[9].value, +new Date(Date.UTC(2014, 0, 1)));
+      });
+
+      var table = {
+        columns: [
+          {
+            name: 'a',
+            type: TYPES.Bit
+          },
+          {
+            name: 'b',
+            type: TYPES.TinyInt
+          },
+          {
+            name: 'c',
+            type: TYPES.SmallInt
+          },
+          {
+            name: 'd',
+            type: TYPES.Int
+          },
+          {
+            name: 'e',
+            type: TYPES.BigInt
+          },
+          {
+            name: 'f',
+            type: TYPES.Real
+          },
+          {
+            name: 'g',
+            type: TYPES.Float
+          },
+          {
+            name: 'h',
+            type: TYPES.VarChar,
+            length: 100
+          },
+          {
+            name: 'i',
+            type: TYPES.NVarChar,
+            length: 100
+          },
+          {
+            name: 'j',
+            type: TYPES.DateTime,
+            length: 100
+          }
+        ],
+        rows: [
+          [
+            false,
+            1,
+            2,
+            3,
+            4,
+            5.5,
+            6.6,
+            'asdf',
+            'asdf',
+            new Date(Date.UTC(2014, 0, 1))
+          ]
+        ]
+      };
+
+      request4.addParameter('tvp', TYPES.TVP, table, {});
+
+      connection.callProcedure(request4);
+    });
   });
-
-  connection.on('end', function(info) {
-    test.done();
-  });
-
-  connection.on(
-    'infoMessage',
-    function(info) {}
-    //console.log("#{info.number} : #{info.message}")
-  );
-
-  connection.on(
-    'debug',
-    function(text) {}
-    //console.log(text)
-  );
-};
+});
