@@ -1,5 +1,4 @@
-require('babel-register');
-require('source-map-support').install({
-  handleUncaughtExceptions: false,
-  hookRequire: true
+require('@babel/register')({
+  extensions: ['.js', '.ts'],
+  plugins: [ 'istanbul' ]
 });
