@@ -1,201 +1,237 @@
-var Login7Payload = require('../../src/login7-payload');
+const { assert } = require('chai');
+const Login7Payload = require('../../src/login7-payload');
 
-exports.create = function(test) {
-  var loginData = {
-    userName: 'user',
-    password: 'pw',
-    appName: 'app',
-    serverName: 'server',
-    language: 'lang',
-    database: 'db',
-    packetSize: 1024,
-    tdsVersion: '7_2'
-  };
+describe('Login7Payload', function() {
+  describe('#toBuffer', function() {
 
-  //start = new Date().getTime()
-  //for c in [1..1000]
-  //  payload = new Login7Payload(loginData)
-  //end = new Date().getTime()
-  //console.log(end - start)
-  var payload = new Login7Payload(loginData);
+    describe('for a login payload with a password', function() {
+      it('generates the expected data', function() {
+        const payload = new Login7Payload({
+          tdsVersion: 0x72090002,
+          packetSize: 1024,
+          clientProgVer: 0,
+          clientPid: 12345,
+          connectionId: 0,
+          clientTimeZone: 120,
+          clientLcid: 0x00000409
+        });
 
-  var expectedLength =
-    4 + // Length
-    32 + // Variable
-    2 +
-    2 +
-    2 * payload.hostname.length +
-    2 +
-    2 +
-    2 * loginData.userName.length +
-    2 +
-    2 +
-    2 * loginData.password.length +
-    2 +
-    2 +
-    2 * loginData.appName.length +
-    2 +
-    2 +
-    2 * loginData.serverName.length +
-    2 +
-    2 +
-    2 * 0 + // Reserved
-    2 +
-    2 +
-    2 * payload.libraryName.length +
-    2 +
-    2 +
-    2 * loginData.language.length +
-    2 +
-    2 +
-    2 * loginData.database.length +
-    payload.clientId.length +
-    2 +
-    2 +
-    2 * payload.sspi.length +
-    2 +
-    2 +
-    2 * payload.attachDbFile.length +
-    2 +
-    2 +
-    2 * payload.changePassword.length +
-    4; // cbSSPILong
+        payload.hostname = 'example.com';
+        payload.userName = 'user';
+        payload.password = 'pw';
+        payload.appName = 'app';
+        payload.serverName = 'server';
+        payload.language = 'lang';
+        payload.database = 'db';
+        payload.libraryName = 'Tedious';
+        payload.attachDbFile = 'c:\\mydbfile.mdf';
+        payload.changePassword = 'new_pw';
 
-  test.strictEqual(payload.data.length, expectedLength);
+        const data = payload.toBuffer();
 
-  var passwordStart = payload.data.readUInt16LE(4 + 32 + 2 * 4);
-  var passwordEnd = passwordStart + 2 * loginData.password.length;
-  var passwordExpected = new Buffer([0xa2, 0xa5, 0xd2, 0xa5]);
-  test.ok(
-    payload.data.slice(passwordStart, passwordEnd).equals(passwordExpected)
-  );
+        const expectedLength =
+          4 + // Length
+          32 + // Variable
+          2 +
+          2 +
+          2 * payload.hostname.length +
+          2 +
+          2 +
+          2 * payload.userName.length +
+          2 +
+          2 +
+          2 * payload.password.length +
+          2 +
+          2 +
+          2 * payload.appName.length +
+          2 +
+          2 +
+          2 * payload.serverName.length +
+          2 +
+          2 +
+          2 * 0 + // Reserved
+          2 +
+          2 +
+          2 * payload.libraryName.length +
+          2 +
+          2 +
+          2 * payload.language.length +
+          2 +
+          2 +
+          2 * payload.database.length +
+          6 + // ClientID
+          2 +
+          2 +
+          2 * 0 + // No SSPI given
+          2 +
+          2 +
+          2 * payload.attachDbFile.length +
+          2 +
+          2 +
+          2 * payload.changePassword.length +
+          4 + // cbSSPILong
+          5; // FeatureExt
 
-  //console.log(payload.toString(''))
+        assert.lengthOf(data, expectedLength);
 
-  test.done();
-};
+        const passwordStart = data.readUInt16LE(4 + 32 + 2 * 4);
+        const passwordEnd = passwordStart + 2 * payload.password.length;
+        const passwordExpected = Buffer.from([0xa2, 0xa5, 0xd2, 0xa5]);
 
-exports.createNTLM = function(test) {
-  var loginData = {
-    userName: 'user',
-    password: 'pw',
-    appName: 'app',
-    serverName: 'server',
-    domain: 'domain',
-    workstation: 'workstation',
-    language: 'lang',
-    database: 'db',
-    packetSize: 1024,
-    tdsVersion: '7_2'
-  };
+        assert.deepEqual(data.slice(passwordStart, passwordEnd), passwordExpected);
+      });
+    });
 
-  var payload = new Login7Payload(loginData);
+    describe('for a login payload with SSPI data', function() {
+      it('generates the expected data', function() {
+        const payload = new Login7Payload({
+          tdsVersion: 0x72090002,
+          packetSize: 1024,
+          clientProgVer: 0,
+          clientPid: 12345,
+          connectionId: 0,
+          clientTimeZone: 120,
+          clientLcid: 0x00000409
+        });
 
-  var expectedLength =
-    4 + // Length
-    32 + // Variable
-    2 +
-    2 +
-    2 * payload.hostname.length +
-    2 +
-    2 +
-    2 * loginData.userName.length +
-    2 +
-    2 +
-    2 * loginData.password.length +
-    2 +
-    2 +
-    2 * loginData.appName.length +
-    2 +
-    2 +
-    2 * loginData.serverName.length +
-    2 +
-    2 +
-    2 * 0 + // Reserved
-    2 +
-    2 +
-    2 * payload.libraryName.length +
-    2 +
-    2 +
-    2 * loginData.language.length +
-    2 +
-    2 +
-    2 * loginData.database.length +
-    payload.clientId.length +
-    2 +
-    2 +
-    payload.ntlmPacket.length + // NTLM
-    2 +
-    2 +
-    2 * payload.attachDbFile.length +
-    2 +
-    2 +
-    2 * payload.changePassword.length +
-    4; // cbSSPILong
+        payload.hostname = 'example.com';
+        payload.appName = 'app';
+        payload.serverName = 'server';
+        payload.language = 'lang';
+        payload.database = 'db';
+        payload.libraryName = 'Tedious';
+        payload.attachDbFile = 'c:\\mydbfile.mdf';
+        payload.changePassword = 'new_pw';
+        payload.sspi = Buffer.from([0xa0, 0xa1, 0xa2, 0xa5, 0xd2, 0xa5, 0xa6, 0xa7, 0xa8, 0xa9]);
 
-  test.strictEqual(payload.data.length, expectedLength);
+        const data = payload.toBuffer();
 
-  var protocolHeader = payload.ntlmPacket.slice(0, 8).toString('utf8');
-  test.strictEqual(protocolHeader, 'NTLMSSP\u0000');
+        var expectedLength =
+          4 + // Length
+          32 + // Variable
+          2 + 2 + (2 * payload.hostname.length) +
+          2 + 2 + (2 * 0) +
+          2 + 2 + (2 * 0) +
+          2 + 2 + (2 * payload.appName.length) +
+          2 + 2 + (2 * payload.serverName.length) +
+          2 + 2 + (2 * 0) + // Reserved
+          2 + 2 + (2 * payload.libraryName.length) +
+          2 + 2 + (2 * payload.language.length) +
+          2 + 2 + (2 * payload.database.length) +
+          6 +
+          2 + 2 + payload.sspi.length + // NTLM
+          2 + 2 + (2 * payload.attachDbFile.length) +
+          2 + 2 + (2 * payload.changePassword.length) +
+          4 + // cbSSPILong
+          5; // FeatureExt
 
-  var workstationName = payload.ntlmPacket
-    .slice(payload.ntlmPacket.length - 17)
-    .toString('ascii')
-    .substr(0, 11);
-  test.strictEqual(workstationName, 'WORKSTATION');
+        assert.lengthOf(data, expectedLength);
+      });
+    });
 
-  var domainName = payload.ntlmPacket
-    .slice(payload.ntlmPacket.length - 6)
-    .toString('ascii');
-  test.strictEqual(domainName, 'DOMAIN');
+    describe('for a login payload with active directory authentication', function() {
+      it('generates the expected data', function() {
+        const payload = new Login7Payload({
+          tdsVersion: 0x72090002,
+          packetSize: 1024,
+          clientProgVer: 0,
+          clientPid: 12345,
+          connectionId: 0,
+          clientTimeZone: 120,
+          clientLcid: 0x00000409
+        });
 
-  var passwordStart = payload.data.readUInt16LE(4 + 32 + 2 * 4);
-  var passwordEnd = passwordStart + 2 * loginData.password.length;
-  var passwordExpected = new Buffer([0xa2, 0xa5, 0xd2, 0xa5]);
-  test.ok(
-    payload.data.slice(passwordStart, passwordEnd).equals(passwordExpected)
-  );
+        payload.hostname = 'example.com';
+        payload.appName = 'app';
+        payload.serverName = 'server';
+        payload.language = 'lang';
+        payload.database = 'db';
+        payload.libraryName = 'Tedious';
+        payload.attachDbFile = 'c:\\mydbfile.mdf';
+        payload.changePassword = 'new_pw';
+        payload.fedAuth = {
+          type: 'ADAL',
+          echo: true,
+          workflow: 'default'
+        };
 
-  test.done();
-};
+        const data = payload.toBuffer();
 
-exports.createSSPI = function(test) {
-  var loginData = {
-    userName: '',
-    password: '',
-    appName: 'app',
-    serverName: 'server',
-    domain: 'domain',
-    workstation: 'workstation',
-    language: 'lang',
-    database: 'db',
-    packetSize: 1024,
-    tdsVersion: '7_2',
-    sspiBlob: new Buffer([0xa0, 0xa1, 0xa2, 0xa5, 0xd2, 0xa5, 0xa6, 0xa7, 0xa8, 0xa9])
-  };
+        var expectedLength =
+          4 + // Length
+          32 + // Fixed data
+          // Variable
+          2 + 2 + (2 * payload.hostname.length) +
+          2 + 2 + 2 * 0 + // Username
+          2 + 2 + 2 * 0 + // Password
+          2 + 2 + (2 * payload.appName.length) +
+          2 + 2 + (2 * payload.serverName.length) +
+          2 + 2 + 4 +
+          2 + 2 + (2 * payload.libraryName.length) +
+          2 + 2 + (2 * payload.language.length) +
+          2 + 2 + (2 * payload.database.length) +
+          6 + // ClientID
+          2 + 2 + (2 * payload.attachDbFile.length) +
+          2 + 2 + (2 * payload.changePassword.length) +
+          4 + // cbSSPILong
+          4 + // Extension offset
+          1 + 1 + 4 + 1 + 1; // Feature ext
 
-  var payload = new Login7Payload(loginData);
+        assert.lengthOf(data, expectedLength);
+      });
+    });
 
-  var expectedLength =
-    4 +                                             // Length
-    32 +                                            // Variable
-    2 + 2 + (2 * payload.hostname.length) +
-    2 + 2 + (2 * loginData.userName.length) +
-    2 + 2 + (2 * loginData.password.length) +
-    2 + 2 + (2 * loginData.appName.length) +
-    2 + 2 + (2 * loginData.serverName.length) +
-    2 + 2 + (2 * 0) +                               // Reserved
-    2 + 2 + (2 * payload.libraryName.length) +
-    2 + 2 + (2 * loginData.language.length) +
-    2 + 2 + (2 * loginData.database.length) +
-    payload.clientId.length +
-    2 + 2 + payload.ntlmPacket.length +             // NTLM
-    2 + 2 + (2 * payload.attachDbFile.length) +
-    2 + 2 + (2 * payload.changePassword.length) +
-    4;                                              // cbSSPILong
+    describe('for a login payload with token based authentication', function() {
+      it('generates the expected data', function() {
+        const token = 'validToken';
 
-  test.strictEqual(payload.data.length, expectedLength);
-  test.strictEqual(payload.ntlmPacket, loginData.sspiBlob);
+        const payload = new Login7Payload({
+          tdsVersion: 0x72090002,
+          packetSize: 1024,
+          clientProgVer: 0,
+          clientPid: 12345,
+          connectionId: 0,
+          clientTimeZone: 120,
+          clientLcid: 0x00000409
+        });
 
-  test.done();
-};
+        payload.hostname = 'example.com';
+        payload.appName = 'app';
+        payload.serverName = 'server';
+        payload.language = 'lang';
+        payload.database = 'db';
+        payload.libraryName = 'Tedious';
+        payload.attachDbFile = 'c:\\mydbfile.mdf';
+        payload.changePassword = 'new_pw';
+        payload.fedAuth = {
+          type: 'SECURITYTOKEN',
+          echo: true,
+          fedAuthToken: token
+        };
+        const data = payload.toBuffer();
+
+        const expectedLength =
+          4 + // Length
+          32 + // Fixed data
+          // Variable
+          2 + 2 + (2 * payload.hostname.length) +
+          2 + 2 + 2 * 0 + // Username
+          2 + 2 + 2 * 0 + // Password
+          2 + 2 + (2 * payload.appName.length) +
+          2 + 2 + (2 * payload.serverName.length) +
+          2 + 2 + 4 +
+          2 + 2 + (2 * payload.libraryName.length) +
+          2 + 2 + (2 * payload.language.length) +
+          2 + 2 + (2 * payload.database.length) +
+          6 + // ClientID
+          2 + 2 + (2 * payload.attachDbFile.length) +
+          2 + 2 + (2 * payload.changePassword.length) +
+          4 + // cbSSPILong
+          4 + // Extension offset
+          1 + 4 + 1 + 4 + 1 + (token.length * 2); // Feature ext
+
+        assert.lengthOf(data, expectedLength);
+      });
+    });
+  });
+});
