@@ -1,51 +1,48 @@
 const Connection = require('../../src/tedious').Connection;
 const ISOLATION_LEVEL = require('../../src/tedious').ISOLATION_LEVEL;
 const TYPES = require('../../src/tedious').TYPES;
+const assert = require('chai').assert;
 
-exports.types = function(test) {
-  test.ok(TYPES);
-  test.ok(TYPES.VarChar);
+describe('tedious-test', function() {
 
-  test.done();
-};
+  it('types', () => {
+    assert.ok(TYPES);
+    assert.ok(TYPES.VarChar);
+  });
 
-exports.isolationLevel = function(test) {
-  test.ok(ISOLATION_LEVEL);
-  test.ok(ISOLATION_LEVEL.READ_UNCOMMITTED);
+  it('isolationLevel', () => {
+    assert.ok(ISOLATION_LEVEL);
+    assert.ok(ISOLATION_LEVEL.READ_UNCOMMITTED);
+  });
 
-  test.done();
-};
+  it('connection', () => {
+    assert.ok(Connection);
+  });
 
-exports.connection = function(test) {
-  test.ok(Connection);
-
-  test.done();
-};
-
-exports.connectionDoesNotModifyPassedConfig = function(test) {
-  var config = {
-    server: 'localhost',
-    userName: 'sa',
-    password: 'sapwd',
-    options: {
-      port: 1234,
-      cryptoCredentialsDetails: {
-        ciphers: 'RC4-MD5'
+  it('connectionDoesNotModifyPassedConfig', () => {
+    const config = {
+      server: 'localhost',
+      userName: 'sa',
+      password: 'sapwd',
+      options: {
+        encrypt: false,
+        port: 1234,
+        cryptoCredentialsDetails: {
+          ciphers: 'DEFAULT'
+        }
       }
-    }
-  };
+    };
 
-  var connection = new Connection(config);
+    const connection = new Connection(config);
 
-  test.notStrictEqual(connection.config, config);
-  test.notStrictEqual(connection.config.options, config.options);
+    assert.notStrictEqual(connection.config, config);
+    assert.notStrictEqual(connection.config.options, config.options);
 
-  // Test that we did not do a deep copy of the cryptoCredentialsDetails,
-  // as we never modify that value inside tedious.
-  test.strictEqual(
-    connection.config.options.cryptoCredentialsDetails,
-    config.options.cryptoCredentialsDetails
-  );
-
-  test.done();
-};
+    // Test that we did not do a deep copy of the cryptoCredentialsDetails,
+    // as we never modify that value inside tedious.
+    assert.strictEqual(
+      connection.config.options.cryptoCredentialsDetails,
+      config.options.cryptoCredentialsDetails
+    );
+  });
+});
