@@ -1,77 +1,74 @@
-var dataTypeByName = require('../../../src/data-type').typeByName;
-var WritableTrackingBuffer = require('../../../src/tracking-buffer/tracking-buffer')
-  .WritableTrackingBuffer;
+const dataTypeByName = require('../../../src/data-type').typeByName;
+const WritableTrackingBuffer = require('../../../src/tracking-buffer/writable-tracking-buffer');
+const TokenStreamParser = require('../../../src/token/stream-parser');
+const assert = require('chai').assert;
 
-var TokenStreamParser = require('../../../src/token/stream-parser');
+describe('Colmetadata Token Parser', () => {
+  it('should int', () => {
+    const numberOfColumns = 1;
+    const userType = 2;
+    const flags = 3;
+    const columnName = 'name';
 
-module.exports.int = function(test) {
-  var numberOfColumns = 1;
-  var userType = 2;
-  var flags = 3;
-  var columnName = 'name';
+    const buffer = new WritableTrackingBuffer(50, 'ucs2');
 
-  var buffer = new WritableTrackingBuffer(50, 'ucs2');
+    buffer.writeUInt8(0x81);
+    buffer.writeUInt16LE(numberOfColumns);
+    buffer.writeUInt32LE(userType);
+    buffer.writeUInt16LE(flags);
+    buffer.writeUInt8(dataTypeByName.Int.id);
+    buffer.writeBVarchar(columnName);
+    // console.log(buffer.data)
 
-  buffer.writeUInt8(0x81);
-  buffer.writeUInt16LE(numberOfColumns);
-  buffer.writeUInt32LE(userType);
-  buffer.writeUInt16LE(flags);
-  buffer.writeUInt8(dataTypeByName.Int.id);
-  buffer.writeBVarchar(columnName);
-  //console.log(buffer.data)
+    const parser = new TokenStreamParser({ token() { } }, {}, {});
+    parser.write(buffer.data);
+    const token = parser.read();
+    // console.log(token)
 
-  var parser = new TokenStreamParser({ token() {} }, {}, {});
-  parser.write(buffer.data);
-  var token = parser.read();
-  //console.log(token)
+    assert.isOk(!token.error);
+    assert.strictEqual(token.columns.length, 1);
+    assert.strictEqual(token.columns[0].userType, 2);
+    assert.strictEqual(token.columns[0].flags, 3);
+    assert.strictEqual(token.columns[0].type.name, 'Int');
+    assert.strictEqual(token.columns[0].colName, 'name');
+  });
 
-  test.ok(!token.error);
-  test.strictEqual(token.columns.length, 1);
-  test.strictEqual(token.columns[0].userType, 2);
-  test.strictEqual(token.columns[0].flags, 3);
-  test.strictEqual(token.columns[0].type.name, 'Int');
-  test.strictEqual(token.columns[0].colName, 'name');
+  it('should varchar', () => {
+    const numberOfColumns = 1;
+    const userType = 2;
+    const flags = 3;
+    const length = 3;
+    const collation = Buffer.from([0x09, 0x04, 0x50, 0x78, 0x9a]);
+    const columnName = 'name';
 
-  test.done();
-};
+    const buffer = new WritableTrackingBuffer(50, 'ucs2');
 
-module.exports.varchar = function(test) {
-  var numberOfColumns = 1;
-  var userType = 2;
-  var flags = 3;
-  var length = 3;
-  var collation = new Buffer([0x09, 0x04, 0x50, 0x78, 0x9a]);
-  var columnName = 'name';
+    buffer.writeUInt8(0x81);
+    buffer.writeUInt16LE(numberOfColumns);
+    buffer.writeUInt32LE(userType);
+    buffer.writeUInt16LE(flags);
+    buffer.writeUInt8(dataTypeByName.VarChar.id);
+    buffer.writeUInt16LE(length);
+    buffer.writeBuffer(collation);
+    buffer.writeBVarchar(columnName);
+    // console.log(buffer)
 
-  var buffer = new WritableTrackingBuffer(50, 'ucs2');
+    const parser = new TokenStreamParser({ token() { } }, {}, {});
+    parser.write(buffer.data);
+    const token = parser.read();
+    // console.log(token)
 
-  buffer.writeUInt8(0x81);
-  buffer.writeUInt16LE(numberOfColumns);
-  buffer.writeUInt32LE(userType);
-  buffer.writeUInt16LE(flags);
-  buffer.writeUInt8(dataTypeByName.VarChar.id);
-  buffer.writeUInt16LE(length);
-  buffer.writeBuffer(collation);
-  buffer.writeBVarchar(columnName);
-  //console.log(buffer)
-
-  var parser = new TokenStreamParser({ token() {} }, {}, {});
-  parser.write(buffer.data);
-  var token = parser.read();
-  //console.log(token)
-
-  test.ok(!token.error);
-  test.strictEqual(token.columns.length, 1);
-  test.strictEqual(token.columns[0].userType, 2);
-  test.strictEqual(token.columns[0].flags, 3);
-  test.strictEqual(token.columns[0].type.name, 'VarChar');
-  test.strictEqual(token.columns[0].collation.lcid, 0x0409);
-  test.strictEqual(token.columns[0].collation.codepage, 'CP1257');
-  test.strictEqual(token.columns[0].collation.flags, 0x57);
-  test.strictEqual(token.columns[0].collation.version, 0x8);
-  test.strictEqual(token.columns[0].collation.sortId, 0x9a);
-  test.strictEqual(token.columns[0].colName, 'name');
-  test.strictEqual(token.columns[0].dataLength, length);
-
-  test.done();
-};
+    assert.isOk(!token.error);
+    assert.strictEqual(token.columns.length, 1);
+    assert.strictEqual(token.columns[0].userType, 2);
+    assert.strictEqual(token.columns[0].flags, 3);
+    assert.strictEqual(token.columns[0].type.name, 'VarChar');
+    assert.strictEqual(token.columns[0].collation.lcid, 0x0409);
+    assert.strictEqual(token.columns[0].collation.codepage, 'CP1257');
+    assert.strictEqual(token.columns[0].collation.flags, 0x57);
+    assert.strictEqual(token.columns[0].collation.version, 0x8);
+    assert.strictEqual(token.columns[0].collation.sortId, 0x9a);
+    assert.strictEqual(token.columns[0].colName, 'name');
+    assert.strictEqual(token.columns[0].dataLength, length);
+  });
+});
