@@ -1,73 +1,70 @@
-var Parser = require('../../../src/token/stream-parser');
-var WritableTrackingBuffer = require('../../../src/tracking-buffer/tracking-buffer')
-  .WritableTrackingBuffer;
+const Parser = require('../../../src/token/stream-parser');
+const WritableTrackingBuffer = require('../../../src/tracking-buffer/writable-tracking-buffer');
+const assert = require('chai').assert;
 
-module.exports.database = function(test) {
-  var oldDb = 'old';
-  var newDb = 'new';
+describe('Env Change Token Parser', () => {
+  it('should write to database', () => {
+    const oldDb = 'old';
+    const newDb = 'new';
 
-  var buffer = new WritableTrackingBuffer(50, 'ucs2');
+    const buffer = new WritableTrackingBuffer(50, 'ucs2');
 
-  buffer.writeUInt8(0xe3);
-  buffer.writeUInt16LE(0); // Length written later
-  buffer.writeUInt8(0x01); // Database
-  buffer.writeBVarchar(newDb);
-  buffer.writeBVarchar(oldDb);
+    buffer.writeUInt8(0xe3);
+    buffer.writeUInt16LE(0); // Length written later
+    buffer.writeUInt8(0x01); // Database
+    buffer.writeBVarchar(newDb);
+    buffer.writeBVarchar(oldDb);
 
-  var data = buffer.data;
-  data.writeUInt16LE(data.length - 3, 1);
+    const data = buffer.data;
+    data.writeUInt16LE(data.length - 3, 1);
 
-  var parser = new Parser({ token() {} }, {}, {});
-  parser.write(data);
-  var token = parser.read();
+    const parser = new Parser({ token() { } }, {}, {});
+    parser.write(data);
+    const token = parser.read();
 
-  test.strictEqual(token.type, 'DATABASE');
-  test.strictEqual(token.oldValue, 'old');
-  test.strictEqual(token.newValue, 'new');
+    assert.strictEqual(token.type, 'DATABASE');
+    assert.strictEqual(token.oldValue, 'old');
+    assert.strictEqual(token.newValue, 'new');
+  });
 
-  test.done();
-};
+  it('should write with correct packet size', () => {
+    const oldSize = '1024';
+    const newSize = '2048';
 
-module.exports.packetSize = function(test) {
-  var oldSize = '1024';
-  var newSize = '2048';
+    const buffer = new WritableTrackingBuffer(50, 'ucs2');
 
-  var buffer = new WritableTrackingBuffer(50, 'ucs2');
+    buffer.writeUInt8(0xe3);
+    buffer.writeUInt16LE(0); // Length written later
+    buffer.writeUInt8(0x04); // Packet size
+    buffer.writeBVarchar(newSize);
+    buffer.writeBVarchar(oldSize);
 
-  buffer.writeUInt8(0xe3);
-  buffer.writeUInt16LE(0); // Length written later
-  buffer.writeUInt8(0x04); // Packet size
-  buffer.writeBVarchar(newSize);
-  buffer.writeBVarchar(oldSize);
+    const data = buffer.data;
+    data.writeUInt16LE(data.length - 3, 1);
 
-  var data = buffer.data;
-  data.writeUInt16LE(data.length - 3, 1);
+    const parser = new Parser({ token() { } }, {}, {});
+    parser.write(data);
+    const token = parser.read();
 
-  var parser = new Parser({ token() {} }, {}, {});
-  parser.write(data);
-  var token = parser.read();
+    assert.strictEqual(token.type, 'PACKET_SIZE');
+    assert.strictEqual(token.oldValue, 1024);
+    assert.strictEqual(token.newValue, 2048);
+  });
 
-  test.strictEqual(token.type, 'PACKET_SIZE');
-  test.strictEqual(token.oldValue, 1024);
-  test.strictEqual(token.newValue, 2048);
+  it('should be of bad type', () => {
+    const buffer = new WritableTrackingBuffer(50, 'ucs2');
 
-  test.done();
-};
+    buffer.writeUInt8(0xe3);
+    buffer.writeUInt16LE(0); // Length written later
+    buffer.writeUInt8(0xff); // Bad type
 
-module.exports.badType = function(test) {
-  var buffer = new WritableTrackingBuffer(50, 'ucs2');
+    const data = buffer.data;
+    data.writeUInt16LE(data.length - 3, 1);
 
-  buffer.writeUInt8(0xe3);
-  buffer.writeUInt16LE(0); // Length written later
-  buffer.writeUInt8(0xff); // Bad type
+    const parser = new Parser({ token() { } }, {}, {});
+    parser.write(data);
+    const token = parser.read();
 
-  var data = buffer.data;
-  data.writeUInt16LE(data.length - 3, 1);
-
-  var parser = new Parser({ token() {} }, {}, {});
-  parser.write(data);
-  var token = parser.read();
-
-  test.strictEqual(token, null);
-  test.done();
-};
+    assert.strictEqual(token, null);
+  });
+});
