@@ -1,22 +1,22 @@
 const TransientErrorLookup = require('../../src/transient-error-lookup').TransientErrorLookup;
+const assert = require('chai').assert;
 
 // This test is simply a set of assertions to ensure any additions to or deletions
 // from the list of transient errors is intentional.
-exports['Connection configuration validation'] = {
-  'transient errors': function(test) {
-    const transientErrorLookup = new TransientErrorLookup(this.config);
-    test.ok(transientErrorLookup.isTransientError(4060));
-    test.ok(transientErrorLookup.isTransientError(10928));
-    test.ok(transientErrorLookup.isTransientError(10929));
-    test.ok(transientErrorLookup.isTransientError(40197));
-    test.ok(transientErrorLookup.isTransientError(40501));
-    test.ok(transientErrorLookup.isTransientError(40613));
-    test.done();
-  },
 
-  'not transient error': function(test) {
-    const transientErrorLookup = new TransientErrorLookup(this.config);
-    test.strictEqual(transientErrorLookup.isTransientError(18456), false);
-    test.done();
-  }
-};
+describe('Connection configuration validation', (done) => {
+  it('transient errors', () => {
+    const transientErrorLookup = new TransientErrorLookup();
+    assert.ok(transientErrorLookup.isTransientError(4060));
+    assert.ok(transientErrorLookup.isTransientError(10928));
+    assert.ok(transientErrorLookup.isTransientError(10929));
+    assert.ok(transientErrorLookup.isTransientError(40197));
+    assert.ok(transientErrorLookup.isTransientError(40501));
+    assert.ok(transientErrorLookup.isTransientError(40613));
+  });
+
+  it('not transient error', () => {
+    const transientErrorLookup = new TransientErrorLookup();
+    assert.strictEqual(transientErrorLookup.isTransientError(18456), false);
+  });
+});
